% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.cnv.r
\name{amp.del}
\alias{amp.del}
\title{Retrieve amplification events from a 'genecnv.obj' generated by 'gene.cnv' function}
\usage{
amp.del(genecnv.obj, logr.cut = 2)
}
\arguments{
\item{genecnv.obj}{an instance of the class 'genecnv' containing gene level copy number info}

\item{logr.cut}{(numeric) the log-ratio cutoff above which genes are considered amplified (e.g 2 = 8 copies for amplification and 0.5 copies for deep deletions, in diploid regions)}
}
\description{
Retrieve amplification events from a 'genecnv.obj' generated by 'gene.cnv' function
}
