% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.freq.plot.r, R/segment.means.r
\name{ave.segmean}
\alias{ave.segmean}
\title{Obtain the average weighted segment mean log2 ratios from each sample within a CNV segmentaton data.frame}
\usage{
ave.segmean(cnv)

ave.segmean(cnv)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' validated by validate.cnv}
}
\value{
(numeric) a vector containing the weighted average logR from segmented data

(numeric) a vector containing the weighted average logR from segmented data
}
\description{
Obtain the average weighted segment mean log2 ratios from each sample within a CNV segmentaton data.frame

Obtain the average weighted segment mean log2 ratios from each sample within a CNV segmentaton data.frame
}
\examples{

cnv <- validate.cnv(segdat_lung_ccle)
ave.segmean(cnv)

cnv <- validate.cnv(segdat_lung_ccle)
ave.segmean(cnv)
}
\keyword{CNV,}
\keyword{segmentation}
