% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.r
\name{bed2chromo.reg}
\alias{bed2chromo.reg}
\title{Transforms a bed format data.frame containing genomic regions into a matrix of n samples versus m defined genomic bins where bins overlapping with bed segments take value = 1}
\usage{
bed2chromo.reg(
  bed,
  bingr = NULL,
  genome.v = "hg19",
  window.size = 10,
  slide.size = 2
)
}
\arguments{
\item{bed}{(data.frame) An data.frame}

\item{bingr}{(S4) a GenomicRanges object containing the}

\item{genome.v}{(hg19 or hg38) reference genome version to generate genoic bins (ignored if bingr is not NULL)}

\item{window.size}{(numeric) size in megabases to generate genomic bins}

\item{slide.size}{(numeric) size in megabases of the sliding genomic window; slide.size must be <= 1}
}
\value{
an instance of the class 'chromo.regs' containing information about shattered regions
}
\description{
Transforms a bed format data.frame containing genomic regions into a matrix of n samples versus m defined genomic bins where bins overlapping with bed segments take value = 1
}
