% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca.ccle.seg.r
\docType{data}
\name{brca.ccle.seg}
\alias{brca.ccle.seg}
\title{CCLE degmentation data from breast tissue cell lines}
\format{An object of class \code{data.frame} with 45597 rows and 6 columns.}
\usage{
data(brca.ccle.seg)
}
\description{
Affy SNP6.0 array based segmentation data
Downloaded from DepMap: https://depmap.org/portal/download/
}
\examples{
data(brca.ccle.seg)
}
\keyword{datasets}
