% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca.ccle.sv.r
\docType{data}
\name{brca.ccle.sv}
\alias{brca.ccle.sv}
\title{CCLE structural variants from breast tissue cell lines}
\format{An object of class \code{data.frame} with 9678 rows and 8 columns.}
\usage{
data(brca.ccle.sv)
}
\description{
Affy SNP6.0 array based segmentation data
Downloaded from DepMap: https://depmap.org/portal/download/
}
\examples{
data(brca.ccle.sv)
}
\keyword{datasets}
