% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk.burden.iqm.r
\docType{class}
\name{break.iqm-class}
\alias{break.iqm-class}
\alias{break.iqm}
\title{Data class break.iqm}
\arguments{
\item{summary}{(data.table): the frequency of gains and losses in each defined genomic bin}

\item{brk.mat}{(numeric): a matrix of genomic bins versus samples}

\item{chrlimits}{(data.frame): a table containing the chromosome limit coordinates and global genomic coordinates}

\item{plot}{(graphical): a recorded plot object}

\item{param}{(list): a list of parametres provided}
}
\value{
an instance of the class 'cnvfreq'
}
\description{
Class to store breakpoint annotations in association with genomic features (e.g. gene loci)
}
