% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.density.r
\docType{class}
\name{breaks-class}
\alias{breaks-class}
\alias{breaks}
\title{Data class breaks}
\arguments{
\item{breaks}{(data.table): the breakpoint info containing data.table, this will be occupied by the CNV segmentation data in the case of cnv.break.annot or SV for sv.break.annot. Unique random string rownames are added to the returned breaks data.frame.}

\item{burden}{(numeric): a vector containing the total number of breakpoints in each sample}

\item{param}{(list): a list of parametres provided}
}
\value{
an instance of the class 'breaks' containing breakpoint and breakpoint burden information
}
\description{
Class to store breakpoint annotations in association with genomic features (e.g. gene loci)
}
