% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk.burden.iqm.r
\name{brk.burden.iqm}
\alias{brk.burden.iqm}
\title{Evaluates the breakpoint burden based on a instance 'breaks' produced by svpluscnv::scv_breaks or svpluscnv::cnv_breaks. 
Breakpoint densities are calculated for each chromosome arm and the inter quantile mean (svpluscnv::IQM) of al chromosome arms is reported for each sample.
A Graphical output is generated indicating every sample's arm burden ordered by their IQM.}
\usage{
brk.burden.iqm(
  brk,
  sample.col = NULL,
  min.arm.size = 2e+07,
  bp.unit = 1e+07,
  genome.v = "hg19",
  chr.lim = NULL,
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{brk}{(breaks) An instance of the class 'breaks' obtained from CNV segmentation data (svpluscnv::cnv.breaks) or Structural Variant calls (svpluscnv::svc.breaks).}

\item{sample.col}{(character) A vector of valid colors. Names must match sample column from 'brk'. If null a gradiant color based on breakpoint burden IQM will be used.}

\item{min.arm.size}{(numeric) minimum size in base pairs for a chromosome arm to be included in the analysis. Size will be calculated based on the 'genome.v' centromere location (excluding centromere bands). Chromosome start and en locations can be provided in 'chr.lim'.}

\item{bp.unit}{(numeric) The genomic size unit in base pairs to report brekpoint densities. This parameter is also used for the y axis of the plot.}

\item{genome.v}{(hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{chr.lim}{(data.frame) 3 column table (chrom, begin, end) indicating the chromosome most distal coordinates with coverage. Also returned by the function svpluscnv::chromosome.limit.coords.}

\item{verbose}{(logical) whether to return internal messages}
}
\value{
an instance of the class 'cnvfreq' and optionally a plot into open device
}
\description{
Evaluates the breakpoint burden based on a instance 'breaks' produced by svpluscnv::scv_breaks or svpluscnv::cnv_breaks. 
Breakpoint densities are calculated for each chromosome arm and the inter quantile mean (svpluscnv::IQM) of al chromosome arms is reported for each sample.
A Graphical output is generated indicating every sample's arm burden ordered by their IQM.
}
\examples{

# initialize CNV data
svc <- validate.svc(nbl_svdat)

# obtain CNV breakpoints
brk <- cnv.breaks(cnv)

brk.burden.iqm(brk)
}
\keyword{burden,}
\keyword{chromosomal}
\keyword{instability}
\keyword{mutational}
\keyword{structural}
\keyword{variants,}
