% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.r
\docType{class}
\name{chromo.regs-class}
\alias{chromo.regs-class}
\alias{chromo.regs}
\title{Class to store shattered regions and information produced by shattered.regions and shattered.regions.cnv functions}
\arguments{
\item{regions.summary}{(list): a list of data.frames sumarizing the information of shattered regions found in each sample}

\item{high.density.regions}{(matrix): a numeric matrix representing high breakpoint density genomic bins in each sample (values 1 = high density break; 0 = normal)}

\item{high.density.regions.hc}{(matrix): a numeric matrix representing high breakpoint density genomic bins in each sample (values 1 = high density break; 0 = normal). 
Only those bins that overlap with high confidence regions defined in regions.summary are set to = 1}

\item{seg.brk.dens}{(matrix): a numeric matrix representing the number of CNV segmentation breakpoints found in at genomic bins in each sample}

\item{sv.brk.dens}{(matrix): a numeric matrix representing the number of SV breakpoints found at genomic bins in each sample}

\item{seg.brk.common.dens}{(matrix): a numeric matrix representing the number of CNV breakpoints colocalizing SV breakpoints found at genomic bins in each sample}

\item{sv.brk.common.dens}{(matrix): a numeric matrix representing the number of SV breakpoints colocalizing CNV breakpoints found at genomic bins in each sample}

\item{segbrk}{(list): on object generated by seg.breaks function}

\item{svbrk}{(list): on object generated by sv.breaks function}

\item{common.brk}{(list): on object generated by match.breaks function}

\item{segdat}{(data.frame): data.frame containing CNV segmentation data as provided in the input arguments}

\item{svdat}{(data.frame): data.frame containing SV calls as provided in the input arguments}

\item{param}{(list): list of configuration parameters provided or set as default}
}
\value{
an instance of the class 'chromo.regs' containing breakpoint mapping onto genes
}
\description{
Class to store shattered regions and information produced by shattered.regions and shattered.regions.cnv functions
}
