% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ.chromo.plot.r
\name{circ.chromo.plot}
\alias{circ.chromo.plot}
\title{Circos plot combining segmentation and SV calls}
\usage{
circ.chromo.plot(
  shatt.regions,
  sample.id,
  lrr.pct = 0.2,
  lrr.max = 4,
  genome.v = "hg19",
  chrlist = NULL
)
}
\arguments{
\item{shatt.regions}{(list) an object returned by "shattered.regions" function}

\item{sample.id}{(character) the id of a sample to be plotted within names(shatt.regions$regions.summ)}

\item{lrr.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents 20 percent fold change}

\item{lrr.max}{(numeric) CNV plot limit}

\item{genome.v}{(hg19 or h38) reference genome version to draw chromosome limits and centromeres}

\item{chrlist}{(character) vector containing chromosomes to plot; by default only chromosomes with shattered regions are ploted}
}
\description{
Circos plot combining segmentation and SV calls
}
\examples{
circ.chromo.plot()
}
\keyword{CNV,}
\keyword{circular}
\keyword{plot}
\keyword{segmentation,}
\keyword{structural}
\keyword{variant,}
\keyword{visualization,}
