% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ.wg.plot.r
\name{circ.wg.plot}
\alias{circ.wg.plot}
\title{Circos plot combining segmentation and SV calls}
\usage{
circ.wg.plot(
  seg,
  sv,
  sample.id = NULL,
  lrr.pct = 0.2,
  lrr.max = 4,
  genome.v = "hg19",
  chr.list = NULL
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with at least 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{sv}{(data.frame) structural variant table including  8 columns: sample, chrom1, pos1, strand1, chrom2, pos2, strand2, svclass}

\item{lrr.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2}

\item{lrr.max}{(numeric) maximum fold change to be plotted (default = 4)}

\item{genome.v}{(hg19 or h38) reference genome version to draw chromosome limits and centromeres}

\item{low.cov}{(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}
}
\description{
Circos plot combining segmentation and SV calls
}
\examples{
read.depth.breaks()
}
\keyword{CNV,}
\keyword{circular}
\keyword{plot}
\keyword{segmentation,}
\keyword{structural}
\keyword{variant,}
\keyword{visualization,}
