% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.cnv.artifact.r
\name{clean.cnv.artifact}
\alias{clean.cnv.artifact}
\title{Obtain a matrix with the weighted average CN per chromosome arm}
\usage{
clean.cnv.artifact(
  seg,
  n.reps = 4,
  seg.size = 5e+06,
  pc.overlap = 0.99,
  fill = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{n.reps}{(numeric) number of samples with identical segment to consider artifact}

\item{seg.size}{(numeric) only smaller segments will be modified in the seg data.frame}

\item{pc.overlap}{(numeric) minimun percentage overlap for a pair of segments to be consider identical}

\item{fill}{(logical) whether to fill gaps from the segmentaed file after filtering artifacts}

\item{verbose}{(logical)}
}
\description{
Obtain a matrix with the weighted average CN per chromosome arm
}
\examples{

## validate input data.frame
seg <- validate.seg(segdat_lung_ccle)

clean.cnv.artifact(seg)
}
\keyword{CNV,}
\keyword{filter}
\keyword{segmentation,}
