% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.blacklist.regions.r
\docType{data}
\name{cnv.blacklist.regions}
\alias{cnv.blacklist.regions}
\title{genomic regions of low coverage for hg19}
\format{An object of class \code{data.frame} with 60 rows and 3 columns.}
\usage{
data(cnv.blacklist.regions)
}
\description{
This repertoir of low coverage regions was based on affy SNP6.0 arrays
}
\examples{
data(cnv.blacklist.regions)
}
\keyword{datasets}
