% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.annot.r
\name{cnv.break.annot}
\alias{cnv.break.annot}
\title{Identify recurrently altered genes}
\usage{
cnv.break.annot(
  seg,
  fc.pct = 0.2,
  genome.v = "hg19",
  genesgr = NULL,
  upstr = 50000,
  dnstr = 50000,
  min.seg.size = NULL,
  min.num.probes = NULL,
  low.cov = NULL,
  clean.brk = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{fc.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2}

\item{genome.v}{(hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{upstr}{(numeric) genomic region size (base pairs) to identify breakpoints overlapping with gene upstream  regions}

\item{dnstr}{(numeric) genomic region size (base pairs) to identify breakpoints overlapping with gene downstream  regions}

\item{min.seg.size}{(numeric) The minimun segment size (in base pairs) to include in the analysis}

\item{min.num.probes}{(numeric) The minimun number of probes per segment to include in the analysis}

\item{low.cov}{(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}

\item{clean.brk}{(numeric) identical CNV segments across multiple samples are removed and replaced by neighbours average}

\item{verbose}{(logical)}

\item{genegr}{(GRanges) a GRanges object containing gene annotations (if not NULL overides genom.v). It must containg 'strand' and a metadata field 'gene_id' with unique values. Seqnames are expected in the format (chr1, chr2, ...)}
}
\value{
an instance of the class 'break.annot' containing breakpoint mapping onto genes
}
\description{
Identify recurrently altered genes
}
\examples{

## validate input data.frame
seg <- validate.seg(segdat_lung_ccle)

cnv.break.annot(seg)
}
\keyword{CNV,}
\keyword{segmentation}
