% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.density.r
\name{cnv.breaks}
\alias{cnv.breaks}
\title{Identify read-depth breakpoints provided a segmentation file}
\usage{
cnv.breaks(
  cnv,
  fc.pct = 0.2,
  break.width = NULL,
  min.cnv.size = NULL,
  min.num.probes = NULL,
  chrlist = NULL,
  low.cov = NULL,
  clean.brk = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cnv}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{fc.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2}

\item{break.width}{(numeric) the maximum distance between a segment end and the subsequent segment start positions beyond which breakpoints are discarded}

\item{min.cnv.size}{(numeric) The minimun segment size (in base pairs) to include in the analysis}

\item{min.num.probes}{(numeric) The minimun number of probes per segment to include in the analysis}

\item{low.cov}{(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}

\item{clean.brk}{(numeric) identical breakpoints across multiple samples tend to be artifacts; remove breaks > N}
}
\description{
Identify read-depth breakpoints provided a segmentation file
}
\examples{

## Obtain breakpoints from segmentation data
cnv <- validate.cnv(segdat_lung_ccle)

cnv.breaks(cnv)
}
\keyword{CNV,}
\keyword{segmentation}
