% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.freq.plot.r
\name{cnv.freq.plot}
\alias{cnv.freq.plot}
\title{Plot CNV frequency across the human genome from a seg file sontaining multiple samples}
\usage{
cnv.freq.plot(
  seg = NULL,
  ch.pct = 0.2,
  g.bin = 1,
  hg = "hg19",
  cex.axis = 1,
  cex.lab = 1,
  axis.line = -1.5,
  label.line = -1.2
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{ch.pct}{(numeric) percentage CNV gain/loss for a segment to be considered changed (i.e. 0.2 = 20 percent change 0.8 < segmean && segmean > 1.2)}

\item{g.bin}{(numeric) size in megabases of the genmome bin to compute break density}

\item{hg}{(hg19 or h38) reference genome version to draw chromosome limits and centromeres}

\item{cex.axis, cex.lab, axis.line, label.line}{plot parameters}
}
\description{
Plot CNV frequency across the human genome from a seg file sontaining multiple samples
}
\examples{
cnv.freq.plot()
}
\keyword{CNV,}
\keyword{segmentation}
