% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.gene.breaks.r
\name{cnv.gene.breaks}
\alias{cnv.gene.breaks}
\title{Identify recurrently altered genes}
\usage{
cnv.gene.breaks(
  seg,
  fc.pct = 0.2,
  genome.v = "hg19",
  geneid = "Symbol",
  maxgap = 1000,
  upstr = 50000,
  min.seg.size = NULL,
  min.num.probes = NULL,
  low.cov = NULL,
  clean.brk = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{chr.lim}{(data.frame) 3 column table (chrom, begin, end) returned by chromosome.limit.coords}

\item{window.size}{(numeric) size in megabases of the genmome bin to compute break density}

\item{slide.size}{(numeric) size in megabases of the sliding genmome window}
}
\description{
Identify recurrently altered genes
}
\examples{
cnv.gene.breaks()
}
\keyword{CNV,}
\keyword{segmentation}
