% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{data}
\name{cnv_blacklist_regions}
\alias{cnv_blacklist_regions}
\title{Low coverage regions}
\format{
An object of class \code{data.frame} with 60 rows and 3 columns.
}
\usage{
cnv_blacklist_regions
}
\description{
Low coverage regions
}
\keyword{CNV}
\keyword{segmentation}
