% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.freq.plot.r
\docType{class}
\name{cnvfreq-class}
\alias{cnvfreq-class}
\alias{cnvfreq}
\title{Data class cnvfreq}
\arguments{
\item{freqsum}{(data.table): the frequency of gains and losses in each defined genomic bin}

\item{chrlimits}{(data.frame): a table containing the chromosome limit coordinates and global genomic coordinates}

\item{bin.mat}{(numeric): a matrix of genomic bins versus samples}

\item{plot}{(graphical): a recorded plot object}

\item{param}{(list): a list of parametres provided}
}
\value{
an instance of the class 'cnvfreq'
}
\description{
Class to store breakpoint annotations in association with genomic features (e.g. gene loci)
}
