% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomString.r
\name{createRandomString}
\alias{createRandomString}
\title{Obtains inter quantile standard deviation for a defined 'x' vector and both lower and upper quantiles}
\usage{
createRandomString(n = 1, strlen = 12, seed = 123456789)
}
\arguments{
\item{n}{character vector length to return}

\item{strlen}{random string length}

\item{seed}{set.seed value}
}
\description{
Obtains inter quantile standard deviation for a defined 'x' vector and both lower and upper quantiles
}
\examples{
createRandomString()
}
\keyword{random}
\keyword{string}
