% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{d3gb.chr.lim}
\alias{d3gb.chr.lim}
\title{A function to order a list of chromosomes}
\usage{
d3gb.chr.lim(genome.v)
}
\arguments{
\item{chrlist}{(character): a vector containing chromosome names (chr1, chr2...chrX,chrY  )}
}
\value{
a character vector of sorted chromosomes
}
\description{
A function to order a list of chromosomes
}
\examples{

d3gb.chr.lim(genome.v="hg19")

}
\keyword{CNV,}
\keyword{genes}
\keyword{segmentation,}
