% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.r
\docType{methods}
\name{extract.bins}
\alias{extract.bins}
\alias{extract.bins,chromo.regs-method}
\title{Return the genomicRanges object  containing the genomic bins}
\usage{
extract.bins(object)

\S4method{extract.bins}{chromo.regs}(object)
}
\arguments{
\item{object}{(chromo.regs) An object of class chromo.regs}
}
\value{
an genomicRanges object with defined genomic bins
}
\description{
Return the genomicRanges object  containing the genomic bins
}
