% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.p.test.r
\name{freq.p.test}
\alias{freq.p.test}
\title{Obtains significance cutoff for the frequency of binary events encoded in a matrix}
\usage{
freq.p.test(mat, method = "fdr", p.cut = 0.05, iter = 100)
}
\arguments{
\item{mat}{(numeric matrix) a binary matrix where columns will be tested for their sum value compared to a permutated matrix}

\item{method}{(character) the method to pass to p.adjust function}

\item{p.cut}{(numeric) the cutoff for multiple hypothesis corrected p.value}

\item{iter}{(numeric) Number of iterations to produce null distribution (note that null size will be iter*ncol(mat))}
}
\description{
Obtains significance cutoff for the frequency of binary events encoded in a matrix
}
\examples{
freq.p.test()
}
\keyword{empirical}
\keyword{p.adjust}
\keyword{p.value,}
