% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.p.test.r
\docType{methods}
\name{freq.threshold}
\alias{freq.threshold}
\alias{freq.threshold,null.freq-method}
\title{Return frequency threshold from null.freq object}
\usage{
freq.threshold(object)

\S4method{freq.threshold}{null.freq}(object)
}
\arguments{
\item{object}{(null.freq) An object of class null.freq}
}
\value{
an instance of the class 'chromo.regs' containing breakpoint mapping onto genes
}
\description{
Return frequency threshold from null.freq object
}
