% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.cnv.r
\name{gene.cnv}
\alias{gene.cnv}
\title{Obtain a matrix with the weighted average CN per chromosome arm}
\usage{
gene.cnv(
  cnv,
  genome.v = "hg19",
  genesgr = NULL,
  chrlist = NULL,
  fill.gaps = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{cnv}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{genome.v}{(hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{genesgr}{(S4) a GenomicRanges object containing gene annotations (if not NULL overides genome.v). It must containg 'strand' and a metadata field 'gene_id' with unique values. Seqnames are expected in the format (chr1, chr2, ...)}

\item{chrlist}{(character) list of chromosomes to include chr1, chr2, etc...}

\item{verbose}{(logical)}

\item{fill}{(logical) whether to fill the gaps in the segmentation file using gap neighbour segmean average as log ratio}
}
\description{
Obtain a matrix with the weighted average CN per chromosome arm
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

gene.cnv(cnv)
}
\keyword{CNV,}
\keyword{genes}
\keyword{segmentation,}
