% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{methods}
\name{gene.symbol.info}
\alias{gene.symbol.info}
\alias{gene.symbol.info,refSeqDat-method}
\title{Return coordinates of an specified gene}
\usage{
gene.symbol.info(object, symbol)

\S4method{gene.symbol.info}{refSeqDat}(object, symbol)
}
\arguments{
\item{object}{(refSeqDat) An object of class refSeqDat containing gene transcript mapping. svpluscnv includes two selfloaded objects: refseq_hg19 & refseq_hg38}

\item{symbol}{(character) a valid HGNC gene symbol included in the refseq object}
}
\value{
A list containing chr, start, end coordinates
}
\description{
Return coordinates of an specified gene
}
