% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.track.view.r
\name{gene.track.view}
\alias{gene.track.view}
\title{Integrated visualization of SVs and SNV in local genomic regions}
\usage{
gene.track.view(
  chr = NULL,
  start = NULL,
  stop = NULL,
  symbol = NULL,
  upstr = NULL,
  dnstr = NULL,
  genome.v = "hg19",
  cex.text = 0.6,
  addtext = TRUE,
  show.chr = "topleft",
  plot = TRUE,
  summary = TRUE,
  ...
)
}
\arguments{
\item{chr}{(character) chromosome (e.g chr9)}

\item{start}{(numeric) genomic coordinate from specified chromosome to start plotting}

\item{symbol}{(character) gene acceoted hgnc symbol to retrieve coordinates and area plotting ()}

\item{upstr}{(numeric) distance upstream specified gene to extend the area plotted}

\item{dnstr}{(numeric) distance downstream specified gene to extend the area plotted}

\item{genome.v}{(hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{addtext}{(logic)}

\item{show.chr}{(character) if not null will show the chromosome name at specified location (Default "topleft")}

\item{plot}{(logic)}

\item{summary}{(logic) whether the function shoud return CNV segment 'segbrk' and SV 'svbrk' breakpoints tabular output}

\item{...}{additional plot parameters from graphics plot function}

\item{end}{(numeric) genomic coordinate from specified chromosome to stop plotting}

\item{addlegend}{(logic)}
}
\description{
Integrated visualization of SVs and SNV in local genomic regions
}
\examples{
plot()
}
\keyword{CNV,}
\keyword{segmentation}
