% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.annot.r
\name{geneBreakOverlap}
\alias{geneBreakOverlap}
\title{Find overlaps between genomic features and breakpoints}
\usage{
geneBreakOverlap(ggr, svgr)
}
\arguments{
\item{ggr}{(S4) a GenomicRanges object containing gene annotations. It is crutial that the genome version 'genesgr' and the input 'sv' are the same. The GRanges object must contain 'strand' and a metadata field 'gene_id' with unique values. Seqnames are expected in the format (chr1, chr2, ...).}

\item{svgr}{(S4) a GenomicRanges object containing SV breakpoint ends. Metadata must contain 'rowid' and 'sampleid' fields. Seqnames are expected in the format (chr1, chr2, ...). Used by 'svc.break.annot' and 'cnv.break.annot'}
}
\value{
a list containing two lists: geneBreaks, geneSamples
}
\description{
Find overlaps between genomic features and breakpoints
}
\keyword{internal}
