% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.cnv.r
\docType{class}
\name{genecnv-class}
\alias{genecnv-class}
\alias{genecnv}
\title{Class to store breakpoint annotations}
\arguments{
\item{cnvmat}{(data.frame): chromosome for the first breakpoint}

\item{genesgr}{(S4): a GenomicRanges object with genomic feature annotations such as gene coordinates}

\item{cnv}{(data.frame): segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{param}{(list):}
}
\value{
an instance of the class 'genecnv' containing gene level copy number info
}
\description{
Class to store breakpoint annotations
}
