% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.r
\name{get.chr.bins}
\alias{get.chr.bins}
\title{Generates a GenomicRanges objact containing genomic bins based on a given bin size. If a cnv (svcnvio) object is provided the chromosome limits 
will be obtaind from mapped regions, otherwise chromosome limits will be obtained from the database (D3GB)}
\usage{
get.chr.bins(cnv = NULL, genome.v = "hg19", window.size = 10, slide.size = 2)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}

\item{genome.v}{(hg19 or hg38) reference genome version to generate genoic bins (ignored if cnv is not NULL)}

\item{window.size}{(numeric) size in megabases to generate genomic bins}

\item{slide.size}{(numeric) size in megabases of the sliding genomic window; slide.size must be <= 1}
}
\value{
an instance of the class 'chromo.regs' containing information about shattered regions
}
\description{
Generates a GenomicRanges objact containing genomic bins based on a given bin size. If a cnv (svcnvio) object is provided the chromosome limits 
will be obtaind from mapped regions, otherwise chromosome limits will be obtained from the database (D3GB)
}
