% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.r
\docType{methods}
\name{hbd.mat}
\alias{hbd.mat}
\alias{hbd.mat,chromo.regs-method}
\title{Return the binary matrix containing high confidence high-breakpoint-densityregion definitions}
\usage{
hbd.mat(object, conf = "hc")

\S4method{hbd.mat}{chromo.regs}(object, conf = "hc")
}
\arguments{
\item{object}{(chromo.regs) An object of class chromo.regs}

\item{conf}{(character) Either "hc" for high confidence HBD or else include all}
}
\value{
an instance of the class 'chromo.regs' containing breakpoint mapping onto genes
}
\description{
Return the binary matrix containing high confidence high-breakpoint-densityregion definitions
}
