% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot.spot.samples.R
\name{hot.spot.samples}
\alias{hot.spot.samples}
\title{Define peak regions and samples in shattered regions hot spots}
\usage{
hot.spot.samples(chromo.regs.obj, freq.cut)
}
\arguments{
\item{chromo.regs.obj}{(chromo.regs) An object of class chromo.regs}

\item{freq.cut}{(numeric) the hot spot threshold above which peaks are defined for sample ID retrieval}
}
\value{
a list comprising two lists: peakRegions, peakRegionsSamples
}
\description{
Define peak regions and samples in shattered regions hot spots
}
