% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{IQM}
\alias{IQM}
\title{Obtains inter quantile average for a defined 'x' vector and both lower and upper quantiles}
\usage{
IQM(x, lowQ = 0.1, upQ = 0.9)
}
\arguments{
\item{x}{numeric vector}

\item{lowQ}{lower quantile}

\item{upQ}{upper quantile}
}
\description{
Obtains inter quantile average for a defined 'x' vector and both lower and upper quantiles
}
\examples{

x <- rnorm(100)
IQM(x)
}
\keyword{interquartile}
\keyword{statistics,}
