% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{map2color}
\alias{map2color}
\title{Obtains inter quantile average for a defined 'x' vector and both lower and upper quantiles}
\usage{
map2color(x, pal, limits = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{pal}{color palette}

\item{limits}{numeric limit fr color mapping}
}
\value{
a color vector graded according to x
}
\description{
Obtains inter quantile average for a defined 'x' vector and both lower and upper quantiles
}
\examples{

x <- rnorm(100)
pal <- colorRampPalette(c("lightblue","white","salmon"))(256)
limits<-c(-1,1)
map2color(x,pal,limits)
}
\keyword{color,}
\keyword{number}
