% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.variant.breaks.r
\name{match.variant.breaks}
\alias{match.variant.breaks}
\title{Identify CNV breakpoints provided a segmentation file}
\usage{
match.variant.breaks(
  seg,
  sv,
  maxgap = 50000,
  fc.pct = 0.2,
  low.cov = NULL,
  clean.brk = NULL,
  verbose = FALSE
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with at least 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{sv}{(data.frame) structural variant table including  8 columns: sample, chrom1, pos1, strand1, chrom2, pos2, strand2, svclass}

\item{verbose}{(logical)}

\item{segment.size}{(numeric) size in megabases of the genmome bin to compute break density}
}
\description{
Identify CNV breakpoints provided a segmentation file
}
\examples{
break.density()
}
\keyword{CNV,}
\keyword{segmentation}
