% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.freq.plot.r, R/segment.means.r
\name{med.segmean}
\alias{med.segmean}
\title{Obtain the median segment mean from a segmentaton file; The median refers to the logR that occupies a center of all segments ordered by their log ratio}
\usage{
med.segmean(cnv)

med.segmean(cnv)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' validated by validate.cnv}
}
\value{
(numeric) a vector containing the median logR value of a segmented data.frame

(numeric) a vector containing the median logR value of a segmented data.frame
}
\description{
Obtain the median segment mean from a segmentaton file; The median refers to the logR that occupies a center of all segments ordered by their log ratio

Obtain the median segment mean from a segmentaton file; The median refers to the logR that occupies a center of all segments ordered by their log ratio
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

med.segmean(cnv)

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

med.segmean(cnv)
}
\keyword{CNV,}
\keyword{segmentation}
