% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge2lists.r
\name{merge2lists}
\alias{merge2lists}
\title{Given two lists with (or without common names) returns a combined list; if common names their values merged and returned as unique}
\usage{
merge2lists(x, y)
}
\arguments{
\item{x}{list 1}

\item{y}{list 2}
}
\value{
a combined list
}
\description{
Given two lists with (or without common names) returns a combined list; if common names their values merged and returned as unique
}
\examples{

x <- sapply(letters[1:10], function(i) sample(1:10)[1:sample(2:10)[1]], simplify=FALSE )
y <- sapply(letters[5:15], function(i) sample(1:10)[1:sample(2:10)[1]], simplify=FALSE )
merge2lists(x,y)
}
\keyword{list}
