% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{data}
\name{nbl_segdat}
\alias{nbl_segdat}
\title{TARGET Neuroblastoma CNV}
\format{
An object of class \code{data.frame} with 17680 rows and 6 columns.
}
\usage{
nbl_segdat
}
\description{
TARGET CNV segmentation: https://target-data.nci.nih.gov/
}
\keyword{CNV}
\keyword{SVs}
\keyword{segmentation,}
