% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{data}
\name{nbl_svdat}
\alias{nbl_svdat}
\title{TARGET Neuroblastoma SVC}
\format{An object of class \code{data.frame} with 7366 rows and 8 columns.}
\usage{
nbl_svdat
}
\description{
TARGET CGI structural variants: https://target-data.nci.nih.gov/
}
\keyword{SVs}
