% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.p.test.r
\docType{class}
\name{null.freq-class}
\alias{null.freq-class}
\alias{null.freq}
\title{Data class null.freq}
\arguments{
\item{freq.cut}{(numeric): the value from observed distribution that satisfies certain p-value cutoff}

\item{pvalues}{(numeric): a vector containing the total number of breakpoints in each sample}

\item{observed}{(numeric): vector of observed distribution}

\item{null}{(numeric): vector of null distribution}

\item{param}{(list): a list of parametres provided}
}
\value{
an instance of the class 'freq.cut'
}
\description{
Class to store observed and null distr. as well as ampirical corrected p-values associated with observed values
}
