% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.regions.plot.r
\name{recurrent.regions.plot}
\alias{recurrent.regions.plot}
\title{Plot CNV frequency across the human genome from a seg file sontaining multiple samples}
\usage{
recurrent.regions.plot(
  shattered.regions,
  conf = "hc",
  hg = "hg19",
  fdr_cut = NULL
)
}
\arguments{
\item{shattered.regions}{shattered.regions object returned by 'shattered.regions' or 'shattered.regions.cnv'}

\item{conf}{(character) either 'hc' for high confidence objects or else all included}

\item{hg}{(hg19 or h38) reference genome version to draw chromosome limits and centromeres}

\item{cex.axis, cex.lab, axis.line, label.line}{plot parameters}
}
\description{
Plot CNV frequency across the human genome from a seg file sontaining multiple samples
}
\examples{
recurrent.regions.plot()
}
\keyword{CNV,}
\keyword{segmentation}
