% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{data}
\name{refseq_hg19}
\alias{refseq_hg19}
\title{refSeq annotations for hg19 and hg38 versions from UCSC (http://genome.ucsc.edu/cgi-bin/hgTables)}
\format{An object of class \code{refSeqDat} of length 1.}
\usage{
refseq_hg19
}
\description{
refSeq annotations for hg19 and hg38 versions from UCSC (http://genome.ucsc.edu/cgi-bin/hgTables)
}
\keyword{exons,}
\keyword{genes,}
\keyword{hg19,}
\keyword{hg38}
\keyword{transcripts,}
