% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{data}
\name{refseq_hg38}
\alias{refseq_hg38}
\title{Reference transcript and exon annotations for hg38}
\format{An object of class \code{refSeqDat} of length 1.}
\usage{
refseq_hg38
}
\description{
refSeq annotations for hg38 version from UCSC (http://genome.ucsc.edu/cgi-bin/hgTables)
}
\keyword{exons}
\keyword{genes,}
\keyword{transcripts,}
