% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{class}
\name{refSeqDat-class}
\alias{refSeqDat-class}
\alias{refSeqDat}
\title{Data class refSeqDat}
\arguments{
\item{data}{(data.table): transcript information}

\item{exonStarts}{(list): every transcript exonic end position}

\item{genome.v}{(character): the genome version encoding transcript data}
}
\value{
an instance of the class 'refSeqDat' containing transcript exonic coordinates
}
\description{
Class to store refseq data from UCSC containing exon level info for known transcripts
}
