% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarc_tcga_data.R
\docType{data}
\name{sarc.seg}
\alias{sarc.seg}
\title{ATCGA segmentation data for sarc tumors}
\format{An object of class \code{data.frame} with 82491 rows and 6 columns.}
\usage{
data(sarc.seg)
}
\description{
Affy SNP6.0 array based segmentation data from hg19 legacy version
Downloaded from GDC:
}
\examples{
data(sarc.seg)
}
\keyword{datasets}
