% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.breaks.r
\name{seg.breaks}
\alias{seg.breaks}
\title{Identify read-depth breakpoints provided a segmentation file}
\usage{
seg.breaks(
  seg,
  fc.pct = 0.2,
  min.seg.size = NULL,
  min.num.probes = NULL,
  low.cov = NULL,
  clean.brk = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{fc.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2}

\item{min.seg.size}{(numeric) The minimun segment size (in base pairs) to include in the analysis}

\item{min.num.probes}{(numeric) The minimun number of probes per segment to include in the analysis}

\item{low.cov}{(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}

\item{clean.brk}{(numeric) identical breakpoints across multiple samples tend to be artifacts; remove breaks > N}
}
\description{
Identify read-depth breakpoints provided a segmentation file
}
\examples{
read.depth.breaks()
}
\keyword{CNV,}
\keyword{segmentation}
