% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.cnv.artifact.r
\name{segment.gap}
\alias{segment.gap}
\title{Fills the gaps in a segmentation data.frame. Chromosome limits are defined for the complete segmentation dataset then segments fill the missing terminal regions. 
The CN log-ratio of the added segments is set to the average of the closest neighbours in each sample.}
\usage{
segment.gap(cnv, chrlist = NULL, verbose = TRUE)
}
\arguments{
\item{cnv}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{chrlist}{(character) list of chromosomes to include chr1, chr2, etc...}

\item{verbose}{(logical)}
}
\description{
Fills the gaps in a segmentation data.frame. Chromosome limits are defined for the complete segmentation dataset then segments fill the missing terminal regions. 
The CN log-ratio of the added segments is set to the average of the closest neighbours in each sample.
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

segment.gap()
}
\keyword{CNV,}
\keyword{segmentation}
