% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.eval.r
\name{shattered.eval}
\alias{shattered.eval}
\title{Evaluate shattered regions}
\usage{
shattered.eval(chromo.regs, disp.cut = 0.05, verbose = TRUE)
}
\arguments{
\item{chromo.regs}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{disp.cut}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2}
}
\description{
Evaluate shattered regions
}
\examples{
shattered.eval()
}
\keyword{chromothripsis,}
\keyword{segmentation,}
\keyword{structural}
\keyword{variants}
