% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.cnv.r
\name{shattered.regions.cnv}
\alias{shattered.regions.cnv}
\title{CNV-only based shattered region detection}
\usage{
shattered.regions.cnv(
  cnv,
  fc.pct = 0.2,
  min.cnv.size = 0,
  min.num.probes = 0,
  low.cov = NULL,
  clean.brk = NULL,
  window.size = 10,
  slide.size = 2,
  num.breaks = 10,
  num.sd = 5,
  dist.iqm.cut = 1e+05,
  verbose = TRUE
)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}

\item{fc.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents 20 percent fold change}

\item{min.cnv.size}{(numeric) The minimun segment size (in base pairs) to include in the analysis}

\item{min.num.probes}{(numeric) The minimun number of probes per segment to include in the analysis}

\item{low.cov}{(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}

\item{clean.brk}{(numeric) inherited from cnv.breaks(); n cutoff for redundant breakpoints to filter out; if NULL, no filter will be applied}

\item{window.size}{(numeric) size in megabases of the genmome bin to compute break density}

\item{slide.size}{(numeric) size in megabases of the sliding genmome window}

\item{num.breaks}{(numeric) size in megabases of the genmome bin to compute break density}

\item{num.sd}{(numeric) size in megabases of the sliding genmome window}

\item{dist.iqm.cut}{(numeric) interquantile average of the distance between breakpoints within a shattered region}

\item{verbose}{(logical)}
}
\value{
an instance of the class 'chromo.regs' containing breakpoint mapping onto genes
}
\description{
Caller for the identification of shattered genomic regions based on CNV breakpoint densities
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

shattered.regions.cnv(cnv)
}
\keyword{CNV,}
\keyword{segmentation}
