% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.r
\name{shattered.regions}
\alias{shattered.regions}
\title{Caller for shattered genomic regions based on breakpoint densities}
\usage{
shattered.regions(
  seg,
  sv,
  fc.pct = 0.2,
  min.seg.size = 0,
  min.num.probes = 0,
  low.cov = NULL,
  window.size = 10,
  slide.size = 2,
  num.seg.breaks = 6,
  num.seg.sd = 5,
  num.sv.breaks = 6,
  num.sv.sd = 5,
  num.common.breaks = 3,
  num.common.sd = 3,
  maxgap = 10000,
  chrlist = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{fc.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2}

\item{min.seg.size}{(numeric) The minimun segment size (in base pairs) to include in the analysis}

\item{min.num.probes}{(numeric) The minimun number of probes per segment to include in the analysis}

\item{low.cov}{(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}

\item{window.size}{(numeric) size in megabases of the genmome bin to compute break density}

\item{slide.size}{(numeric) size in megabases of the sliding genmome window}

\item{num.breaks}{(numeric) size in megabases of the genmome bin to compute break density}

\item{num.sd}{(numeric) size in megabases of the sliding genmome window}
}
\description{
Caller for shattered genomic regions based on breakpoint densities
}
\examples{
shattered.regions.cnv()
}
\keyword{CNV,}
\keyword{segmentation}
