% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.annot.r
\name{sv.break.annot}
\alias{sv.break.annot}
\title{Identify recurrently altered genes by strutural variants}
\usage{
sv.break.annot(
  sv,
  genome.v = "hg19",
  genesgr = NULL,
  upstr = 50000,
  dnstr = 50000,
  sv.seg.size = 2e+05,
  verbose = TRUE
)
}
\arguments{
\item{sv}{(data.frame) structural variant table including  8 columns: sample, chrom1, pos1, strand1, chrom2, pos2, strand2, svclass}

\item{genome.v}{(character): either 'hg19' or 'hg38' accepted; reference genome version to retrieve gene annotations}

\item{genesgr}{(S4) a GenomicRanges object containing gene annotations (if not NULL overides genome.v). It must containg 'strand' and a metadata field 'gene_id' with unique values. Seqnames are expected in the format (chr1, chr2, ...)}

\item{upstr}{(numeric) upstream region size (base pairs) to identify breakpoints in gene upstream  regions}

\item{dnstr}{(numeric) upstream region size (base pairs) to identify breakpoints in gene downstream  regions}

\item{sv.seg.size}{(numeric) segmental variants (DEL, DUP, INV or INS) size limit; bigger sigments are considered translocations, in that case, only the breakpoint}
}
\value{
an instance of the class 'break.annot' containing breakpoint mapping onto genes
}
\description{
Identify recurrently altered genes by strutural variants
}
\examples{

## Obtain breakpoints from SV calls data
sv <- validate.sv(svdat_lung_ccle)

sv.break.annot(sv, genome.v="hg19")
}
\keyword{Structural}
\keyword{annotation}
\keyword{variants,}
