% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.breaks.r
\name{sv.breaks}
\alias{sv.breaks}
\title{Identify SV call breakpoints and filter low cov regions if provided}
\usage{
sv.breaks(sv, low.cov = NULL)
}
\arguments{
\item{low.cov}{(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}

\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\description{
Identify SV call breakpoints and filter low cov regions if provided
}
\examples{
read.depth.breaks()
}
\keyword{CNV,}
\keyword{segmentation}
