% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.model.view.r
\name{sv.model.view}
\alias{sv.model.view}
\title{Integrated visualization of SVs and SNV in local genomic regions}
\usage{
sv.model.view(
  sv,
  seg,
  chr,
  start,
  stop,
  cnvlim = c(-2, 2),
  addlegend = "both",
  interval = NULL,
  addtext = NULL,
  text.size = 0.8,
  ...
)
}
\arguments{
\item{sv}{(data.frame) structural variant table including  8 columns: sample, chrom1, pos1, strand1, chrom2, pos2, strand2, svclass}

\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{chr}{(character) chromosome (e.g chr9)}

\item{start}{(numeric) genomic coordinate from specified chromosome to start plotting}

\item{cnvlim}{(numeric) limits for color coding of background CNV log-ratios}

\item{addlegend}{(character) one of 'sv' (show SV type legend), 'cnv' (show CNV background color legend) or 'both'.}

\item{addtext}{(character) a vector indicating what SV types should include text labels indicating brakpoint partners genomic locations. 
The labeels are only added to point tbreakpoint locations outside the plot area. (e.g. c("TRA","INV") )}

\item{text.size}{(numeric) point size of the text}

\item{...}{additional plot parameters from graphics plot function}

\item{end}{(numeric) genomic coordinate from specified chromosome to stop plotting}
}
\description{
Integrated visualization of SVs and SNV in local genomic regions
}
\examples{
plot()
}
\keyword{CNV,}
\keyword{segmentation}
