% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.density.r
\name{svc.breaks}
\alias{svc.breaks}
\title{Identify SV call breakpoints and filter low cov regions if provided}
\usage{
svc.breaks(svc, low.cov = NULL)
}
\arguments{
\item{svc}{(data.frame) structural variant table including  8 columns: sample, chrom1, pos1, strand1, chrom2, pos2, strand2, svclass}

\item{low.cov}{(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}
}
\description{
Identify SV call breakpoints and filter low cov regions if provided
}
\examples{

## Obtain breakpoints from SV calls data
svc <- validate.svc(svdat_lung_ccle)

svc.breaks(svc)
}
\keyword{Structural}
\keyword{variants}
