% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.data.r
\docType{class}
\name{svcnvio-class}
\alias{svcnvio-class}
\alias{svcnvio}
\title{Class to store CNV segmentation data}
\arguments{
\item{df}{(data.frame): cnv or svc data.frames to be validated by 'validate.cnv' or 'validate.svc' respectivelly}
}
\value{
an instance of the class 'CNV' containing segmentation table with validated format for other svpluscnv operations;  A unique id (uid) column is also added
}
\description{
Class to store CNV segmentation data
}
\seealso{
Additional data format information in the man pages of validate.cnv and validate.svc
}
