% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{data}
\name{svdat_lung_ccle}
\alias{svdat_lung_ccle}
\title{Lung CCLE SVC data}
\format{
An object of class \code{data.frame} with 23040 rows and 8 columns.
}
\usage{
svdat_lung_ccle
}
\description{
CCLE translocation data from LUNG tissue cell lines (DepMap): https://depmap.org/portal/download/
}
\keyword{SVs}
