% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.data.r
\name{validate.cnv}
\alias{validate.cnv}
\title{This function calculates the percent genome changed by CNV}
\usage{
validate.cnv(cnv)
}
\arguments{
\item{cnv}{(data.frame) segmentation data with at least 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\description{
This function calculates the percent genome changed by CNV
}
\examples{
validate.cnv()
}
\keyword{CNV,}
\keyword{segmentation}
