% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.seg.r
\name{validate.seg}
\alias{validate.seg}
\title{This function calculates the percent genome changed by CNV}
\usage{
validate.seg(seg)
}
\arguments{
\item{seg}{(data.frame) segmentation data with at least 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\description{
This function calculates the percent genome changed by CNV
}
\examples{
validate.seg()
}
\keyword{CNV,}
\keyword{segmentation}
