% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.cnv.r
\name{amp.del}
\alias{amp.del}
\title{Amplifications and deletions}
\usage{
amp.del(genecnv.obj, logr.cut = 2)
}
\arguments{
\item{genecnv.obj}{(genecnv) an instance of the class 'genecnv' containing gene level copy number info}

\item{logr.cut}{(numeric) the log-ratio cutoff above which genes are considered amplified (e.g 2 = 8 copies for amplification and 0.5 copies for deep deletions, in diploid regions)}
}
\value{
(list) A list of lists including amplified.list, amplified.rank, deepdel.list and deepdel.rank
}
\description{
Retrieve amplification and deletion events from a 'genecnv.obj' generated by 'gene.cnv' function
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

genecnv.obj <- gene.cnv(cnv)

geneampdel <- amp.del(genecnv.obj, logr.cut = 2)
lapply(geneampdel,head)
}
\keyword{CNV,}
\keyword{genes}
\keyword{segmentation,}
