% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.means.r
\name{ave.segmean}
\alias{ave.segmean}
\title{Average sample CNV}
\usage{
ave.segmean(cnv)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}
}
\value{
(numeric) a vector containing the weighted average logR from segmented data
}
\description{
Obtain the weighted average segment mean log2 ratios from each sample within a CNV segmentaton data.frame
}
\examples{

## validate input CNV data.frames
cnv <- validate.cnv(segdat_lung_ccle)

ave_seg_mean <- ave.segmean(cnv)
head(ave_seg_mean)
}
\keyword{CNV,}
\keyword{segmentation}
