% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.means.r
\name{ave.segmean}
\alias{ave.segmean}
\title{Obtain the average weighted segment mean log2 ratios from each sample within a CNV segmentaton data.frame}
\usage{
ave.segmean(cnv)
}
\arguments{
\item{cnv}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\description{
Obtain the average weighted segment mean log2 ratios from each sample within a CNV segmentaton data.frame
}
\examples{

cnv <- validate.cnv(segdat_lung_ccle)
ave.segmean(cnv)
}
\keyword{CNV,}
\keyword{segmentation}
