% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ave.segmean.r
\name{ave.segmean}
\alias{ave.segmean}
\title{Obtain the average weighted segment mean from a segmentaton file}
\usage{
ave.segmean(seg)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\description{
Obtain the average weighted segment mean from a segmentaton file
}
\examples{
ave.segmean()
}
\keyword{CNV,}
\keyword{segmentation}
