% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.annot.r
\docType{class}
\name{break.annot-class}
\alias{break.annot-class}
\alias{break.annot}
\title{Class to store breakpoint annotations}
\arguments{
\item{input}{(data.frame): chromosome for the first breakpoint}

\item{genesgr}{(GRanges): a GRanges object with genomic segments (e.g. genes)}

\item{disruptSamples}{(list): chromosome for the second breakpoint}

\item{disruptBreaks}{(list): position for the second breakpoint}

\item{upstreamSamples}{(list):}

\item{upstreamBreaks}{(list):}

\item{param}{(list):}
}
\value{
an instance of the class 'break.annot' containing breakpoint mapping onto genes
}
\description{
Class to store breakpoint annotations
}
