% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.density.r
\name{break.density}
\alias{break.density}
\title{Identify CNV breakpoints provided a segmentation file}
\usage{
break.density(
  brk,
  chr.lim,
  window.size = 10,
  slide.size = 2,
  chrlist = NULL,
  verbose = TRUE
)
}
\arguments{
\item{brk}{(data.frame) 3 column table (sample, chrom, position) i.e returned by read.depth.breaks}

\item{chr.lim}{(data.frame) 3 column table (chrom, begin, end) returned by chromosome.limit.coords}

\item{window.size}{(numeric) size in megabases of the genmome bin to compute break density}

\item{slide.size}{(numeric) size in megabases of the sliding genmome window}

\item{chrlist}{(character) vector containing chromosomes to consider (e.g. c("chr1", "chr2", "chr3", ...))}

\item{verbose}{(logical)}
}
\description{
Identify CNV breakpoints provided a segmentation file
}
\examples{

seg <- validate.seg(segdat_lung_ccle)
brk <- seg.breaks(seg)
chr.lim <- chromosome.limit.coords(seg)
break.density(brk,chr.lim)
}
\keyword{CNV,}
\keyword{segmentation}
