% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.density.r
\name{break.density}
\alias{break.density}
\title{Identify CNV breakpoints provided a segmentation file}
\usage{
break.density(
  breaks,
  chr.lim,
  window.size = 10,
  slide.size = 2,
  chrlist = NULL,
  verbose = TRUE
)
}
\arguments{
\item{breaks}{(data.frame) 3 column table (sample, chrom, position) i.e returned by read.depth.breaks}

\item{chr.lim}{(data.frame) 3 column table (chrom, begin, end) returned by chromosome.limit.coords}

\item{window.size}{(numeric) size in megabases of the genmome bin to compute break density}

\item{slide.size}{(numeric) size in megabases of the sliding genmome window}
}
\description{
Identify CNV breakpoints provided a segmentation file
}
\examples{
break.density()
}
\keyword{CNV,}
\keyword{segmentation}
