% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.density.r
\name{break.density}
\alias{break.density}
\title{Generating a genomic map based on a defined bin size and sliding window and counts the number of breakpoints mapped onto each bin. This function is used internally by svpluscnv::shattered.regions and svpluscnv::shattered.regions.cnv}
\usage{
break.density(
  brk,
  chr.lim = NULL,
  genome.v = "hg19",
  window.size = 10,
  slide.size = 2,
  verbose = TRUE
)
}
\arguments{
\item{brk}{(breaks) An instance of the class 'breaks' obtained from CNV segmentation data (svpluscnv::cnv.breaks) or Structural Variant calls (svpluscnv::svc.breaks).}

\item{chr.lim}{(data.frame) 3 column table (chrom, begin, end) indicating the chromosome most distal coordinates with coverage. Also returned by the function svpluscnv::chromosome.limit.coords.}

\item{genome.v}{(hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{window.size}{(numeric) size in megabases of the genmome bin onto which breakpoints will be mapped}

\item{slide.size}{(numeric) size in megabases of the sliding genomic window; if slide.size < window.size the genomic bins will overlap}

\item{verbose}{(logical)}
}
\value{
a matrix of samples (rows) and genomic bins (cols) qith the number of breakpoints mapped in heach cell
}
\description{
Generating a genomic map based on a defined bin size and sliding window and counts the number of breakpoints mapped onto each bin. This function is used internally by svpluscnv::shattered.regions and svpluscnv::shattered.regions.cnv
}
\examples{

cnv <- validate.cnv(segdat_lung_ccle)
brk <- cnv.breaks(cnv)
chr.lim <- chromosome.limit.coords(cnv)
break.density(brk,chr.lim)
}
\keyword{CNV,}
\keyword{segmentation}
