% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.arm.cnv.r
\name{chr.arm.cnv}
\alias{chr.arm.cnv}
\title{Chromosome arm mean CNV}
\usage{
chr.arm.cnv(cnv, genome.v = "hg19", verbose = FALSE)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' validated by validate.cnv}

\item{genome.v}{(character) (hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{verbose}{(logical) whether to return internal messages}
}
\value{
a matrix of chromosome arms (rows) versus samples (cols) with average segment logRs per cell
}
\description{
Obtains a matrix with the weighted average CN per chromosome arm
}
\examples{

# initialize CNV data
cnv <- validate.cnv(segdat_lung_ccle)

arm_mat <- chr.arm.cnv(cnv, genome.v="hg19")
dim(arm_mat)
}
\keyword{CNV,}
\keyword{arm}
\keyword{chromosome}
\keyword{segmentation,}
