% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.data.r
\name{chr.sort}
\alias{chr.sort}
\title{A function to order a list of chromosomes}
\usage{
chr.sort(chrlist)
}
\arguments{
\item{chrlist}{(character): a vector containing chromosome names (chr1, chr2...chrX,chrY  )}
}
\value{
a character vector of sorted chromosomes
}
\description{
A function to order a list of chromosomes
}
\examples{

chr.sort(chrlist)
}
\keyword{CNV,}
\keyword{genes}
\keyword{segmentation,}
