% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.r
\docType{class}
\name{chromo.regs-class}
\alias{chromo.regs-class}
\alias{chromo.regs}
\title{Data class chromo.regs}
\arguments{
\item{regions.summary}{(list): a list of data.frames sumarizing the information of shattered regions found in each sample}

\item{high.density.regions}{(matrix): a numeric matrix representing high breakpoint density genomic bins in each sample (values 1 = high density break; 0 = normal)}

\item{high.density.regions.hc}{(matrix): a numeric matrix representing high breakpoint density genomic bins in each sample (values 1 = high density break; 0 = normal). 
Only those bins that overlap with high confidence regions defined in regions.summary are set to = 1}

\item{cnv.brk.dens}{(matrix): a numeric matrix representing the number of CNV segmentation breakpoints found in at genomic bins in each sample}

\item{svc.brk.dens}{(matrix): a numeric matrix representing the number of SV breakpoints found at genomic bins in each sample}

\item{cnv.brk.common.dens}{(matrix): a numeric matrix representing the number of CNV breakpoints colocalizing SV breakpoints found at genomic bins in each sample}

\item{svc.brk.common.dens}{(matrix): a numeric matrix representing the number of SV breakpoints colocalizing CNV breakpoints found at genomic bins in each sample}

\item{cnvbrk}{(S4): on object generated by cnv.breaks function}

\item{svcbrk}{(S4): on object generated by svc.breaks function}

\item{common.brk}{(list): on object generated by match.breaks function}

\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' validated by validate.cnv}

\item{svc}{(S4) an object of class svcnvio containing data type 'svc' validated by validate.svc}

\item{param}{(list): list of configuration parameters provided or set as default}
}
\value{
an instance of the class 'chromo.regs' containing breakpoint mapping onto genes
}
\description{
Class to store shattered regions and information produced by shattered.regions and shattered.regions.cnv functions
}
