% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.data.r
\name{chromosome.limit.coords}
\alias{chromosome.limit.coords}
\title{Chromosome limit map}
\usage{
chromosome.limit.coords(cnv)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}
}
\value{
data.table indicating start and end mapped positions of each chromosome
}
\description{
Obtain chromosome start and end positions based on mapped regions from CNV segmentation data
}
\examples{

## validate input data.frame
cnv <- validate.cnv(segdat_lung_ccle)

chr.lim <- chromosome.limit.coords(cnv)
}
\keyword{CNV,}
\keyword{mapping}
\keyword{segmentation,}
