% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome.limit.coords.r
\name{chromosome.limit.coords}
\alias{chromosome.limit.coords}
\title{This function mapps defines mapping limmit regions based on a segmentation file}
\usage{
chromosome.limit.coords(seg)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\description{
This function mapps defines mapping limmit regions based on a segmentation file
}
\examples{

## validate input data.frame
seg <- validate.seg(segdat_lung_ccle)

chr.lim <- chromosome.limit.coords(seg)
}
\keyword{CNV,}
\keyword{mapping}
\keyword{segmentation,}
