% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ.chromo.plot.r
\name{circ.chromo.plot}
\alias{circ.chromo.plot}
\title{Circos plot combining segmentation and SV calls}
\usage{
circ.chromo.plot(
  chromo.regs.obj,
  sample.id,
  genome.v = "hg19",
  lrr.pct = 0.2,
  lrr.max = 4,
  chrlist = NULL
)
}
\arguments{
\item{chromo.regs.obj}{(chromo.regs) An object of class chromo.regs}

\item{sample.id}{(character) the id of a sample to be plotted within names(chromo.regs.obj@regions.summ)}

\item{genome.v}{(hg19 or h38) reference genome version to draw chromosome limits and centromeres}

\item{lrr.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents 20 percent fold change}

\item{lrr.max}{(numeric) CNV plot limit}

\item{chrlist}{(character) vector containing chromosomes to plot; by default only chromosomes with shattered regions are ploted}
}
\description{
Circos plot combining segmentation and SV calls
}
\examples{

require(circlize)

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)
svc <- validate.svc(svdat_lung_ccle)

## obtain shattered regions
shatt.regions <- shattered.regions(cnv,svc)

id <-  names(chromo.regs.obj@regions.summary)[1]
circ.chromo.plot(shatt.regions, sample.id = id)
}
\keyword{CNV,}
\keyword{circular}
\keyword{plot}
\keyword{segmentation,}
\keyword{structural}
\keyword{variant,}
\keyword{visualization,}
