% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.plot.r
\name{circ.wg.plot}
\alias{circ.wg.plot}
\title{Circular visualization CNV and SVC}
\usage{
circ.wg.plot(
  cnv,
  svc,
  sample.id = NULL,
  genome.v = "hg19",
  lrr.pct = 0.2,
  lrr.max = 4,
  chrlist = NULL,
  add.cnv.legend = "topleft",
  add.svc.legend = "toprigh",
  ...
)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}

\item{svc}{(S4) an object of class svcnvio containing data type 'svc' initialized by validate.svc}

\item{sample.id}{(character) the id of the sample to be plotted}

\item{genome.v}{(character) (hg19 or h38) reference genome version to draw chromosome limits and centromeres}

\item{lrr.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2}

\item{lrr.max}{(numeric) maximum CNV to be plotted}

\item{chrlist}{(character) vector containing chromosomes to plot; by default all chromosomes plotted}

\item{add.cnv.legend}{(x,y or coordinates) the position parameter passed to legend to plot CNV (outer tracks) description}

\item{add.svc.legend}{(x,y or coordinates) the position parameter passed to legend to plot SVC (central track) description}
}
\value{
circos plot into open device
}
\description{
Produces a circos plot combining CNV and SVC of the whole genome
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)
svc <- validate.svc(svdat_lung_ccle)

## select a random sample id
id <- "A549_LUNG"

circ.wg.plot(cnv, svc, sample.id=id)
}
\keyword{CNV,}
\keyword{circular}
\keyword{plot}
\keyword{segmentation,}
\keyword{structural}
\keyword{variant,}
\keyword{visualization,}
