% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.cnv.artifact.r
\name{clean.cnv.artifact}
\alias{clean.cnv.artifact}
\title{CNV artifact detection and filtering}
\usage{
clean.cnv.artifact(
  cnv,
  n.reps = 4,
  cnv.size = 2e+06,
  pc.overlap = 0.99,
  fill.gaps = TRUE,
  minsize = 5000,
  verbose = TRUE
)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' validated by validate.cnv}

\item{n.reps}{(numeric) number of samples with identical segment to consider artifact}

\item{cnv.size}{(numeric) only smaller segments will be modified in the cnv data.frame}

\item{pc.overlap}{(numeric) minimun percentage overlap for a pair of segments to be consider identical}

\item{fill.gaps}{(logical) whether to fill gaps from the segmentaed file after filtering artifacts}

\item{minsize}{(numeric) the minimum gap size required to fill the gap. Only used if 'fill.gaps=TRUE'}

\item{verbose}{(logical) whether to print internal messages}
}
\value{
a data.frame containing CNV data
}
\description{
Detects identical or near-identical CNV segments across multiple samples susceptible of representing common variants or technical artifacts. Then those segments CNV log-ratio is replaced by the flanking segments average
}
\examples{

## validate input data.frame
cnv <- validate.cnv(segdat_lung_ccle)

cnvcl <- clean.cnv.artifact(cnv)
cnvcl
}
\keyword{CNV,}
\keyword{filter}
\keyword{segmentation,}
