% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.cnv.artifact.r
\name{clean.cnv.artifact}
\alias{clean.cnv.artifact}
\title{Obtain a matrix with the weighted average CN per chromosome arm}
\usage{
clean.cnv.artifact(
  seg,
  n.reps = 4,
  pc.cut = 0,
  seg.size = 5e+06,
  pc.overlap = 0.99,
  fill = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{fill}{(logical) whether to fill gaps from the segmentaed file after filtering artifacts}

\item{verbose}{(logical)}

\item{reps}{(numeric) number of samples with identical segment to consider artifact}
}
\description{
Obtain a matrix with the weighted average CN per chromosome arm
}
\examples{
clean.cnv.artifact()
}
\keyword{CNV,}
\keyword{filter}
\keyword{segmentation,}
