% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarc_tcga_data.R
\docType{data}
\name{cnv.blacklist.regions}
\alias{cnv.blacklist.regions}
\title{data(cnv.blacklist.regions)}
\format{An object of class \code{data.frame} with 60 rows and 3 columns.}
\usage{
cnv.blacklist.regions
}
\description{
data(cnv.blacklist.regions)
}
\keyword{datasets}
