% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.density.r
\name{cnv.breaks}
\alias{cnv.breaks}
\title{Identify CNV breakpoints}
\usage{
cnv.breaks(
  cnv,
  fc.pct = 0.2,
  break.width = 10000,
  min.cnv.size = NULL,
  min.num.probes = NULL,
  chrlist = NULL,
  low.cov = NULL,
  clean.brk = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}

\item{fc.pct}{(numeric) copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2}

\item{break.width}{(numeric) the maximum distance between a segment end and the subsequent segment start positions beyond which breakpoints are discarded}

\item{min.cnv.size}{(numeric) The minimun segment size (in base pairs) to include in the analysis}

\item{min.num.probes}{(numeric) The minimun number of probes per segment to include in the analysis}

\item{chrlist}{(character) list of chromosomes to include chr1, chr2, etc...}

\item{low.cov}{(data.frame) a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}

\item{clean.brk}{(numeric) identical breakpoints across multiple samples tend to be artifacts; remove breaks > N}

\item{verbose}{(logical) whether to return}
}
\value{
an instance of the class 'breaks' containing breakpoint and breakpoint burden information
}
\description{
Identify CNV breakpoints filtered by the change in copy number log-ratio between contiguous segments
}
\examples{

# initialized CNV data
cnv <- validate.cnv(segdat_lung_ccle)

cnv.breaks(cnv)

}
\keyword{CNV,}
\keyword{segmentation}
