% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.freq.plot.r
\name{cnv.freq}
\alias{cnv.freq}
\title{CNV frequency map}
\usage{
cnv.freq(
  cnv,
  fc.pct = 0.2,
  genome.v = "hg19",
  ploidy = FALSE,
  g.bin = 1,
  sampleids = NULL,
  cex.axis = 1,
  cex.lab = 1,
  label.line = -1.2,
  plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}

\item{fc.pct}{(numeric) percentage CNV gain/loss for a segment to be considered changed (i.e. 0.2 = 20 percent change 0.8 < segmean && segmean > 1.2)}

\item{genome.v}{(character) (hg19 or h38) reference genome version to draw chromosome limits and centromeres}

\item{ploidy}{(logical) whether to apply ploidy correction; the function med.segmean will be used to obtain each sample's ploidy logR then this value substracted to each sample's logR values}

\item{g.bin}{(numeric) size in megabases of the genmome bin to compute break density}

\item{sampleids}{(character) vector containing list of samples to include in plot. if set to NULL, all samples in the input will be used}

\item{cex.axis, cex.lab, label.line}{(numeric) plot parameters}

\item{plot}{(logical) whether produce a graphical output}

\item{verbose}{(logical) whether to return internal messages}
}
\value{
an instance of the class 'cnvfreq' and optionally a plot into open device
}
\description{
Creates a map of CNVs using genome binning and plots CNV frequency across the genome. This function optionally returns text, graphical or both outputs.
}
\examples{

## validate input data.frame
cnv <- validate.cnv(nbl_segdat)

cnv.freq(cnv, genome.v = "hg19")
}
\keyword{CNV,}
\keyword{plot}
\keyword{segmentation,}
