% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{createRandomString}
\alias{createRandomString}
\title{Generates n unique random character strings of a given length}
\usage{
createRandomString(n = 1, strlen = 10)
}
\arguments{
\item{n}{the number of unique random strings to return}

\item{strlen}{random string length}
}
\value{
a vector of unique random character strings
}
\description{
Generates n unique random character strings of a given length
}
\examples{
# for reproducibility make sure to set the seed
set.seed(123456789)
createRandomString(1, 10)
}
\keyword{random}
\keyword{string}
