% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomString.r
\name{createRandomString}
\alias{createRandomString}
\title{Generates n unique random character strings of a given length}
\usage{
createRandomString(n = 1, strlen = 12, seed = 123456789)
}
\arguments{
\item{n}{character vector length to return}

\item{strlen}{random string length}

\item{seed}{set.seed value}
}
\description{
Generates n unique random character strings of a given length
}
\examples{

createRandomString(n=10 strlen=12)
}
\keyword{random}
\keyword{string}
