% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{d3gb.chr.lim}
\alias{d3gb.chr.lim}
\title{Chromosome start and end}
\usage{
d3gb.chr.lim(genome.v)
}
\arguments{
\item{genome.v}{(character) reference genome version to retrieve gene annotations (hg19 or GRCh37 and hg38 or GRCh38)}
}
\value{
(data.table) a table containing start and end positions for each chromosome
}
\description{
Obtains a chromosome start and end positions from a reference genome version
}
\examples{

d3gb.chr.lim(genome.v="hg19")

}
\keyword{CNV,}
\keyword{genes}
\keyword{segmentation,}
