% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.p.test.r
\name{freq.p.test}
\alias{freq.p.test}
\title{Frequency hot spot detection
 
Obtains significance cutoff for the frequency of binary events encoded in a matrix such as that generated by shattered.regions and shattered.regions.cnv algorithms}
\usage{
freq.p.test(
  mat,
  method = "fdr",
  p.cut = 0.05,
  iter = 100,
  zerofreq = TRUE,
  plot = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{mat}{(numeric matrix) a binary matrix where columns will be tested for their sum value compared to a permutated matrix}

\item{method}{(character) the method to pass to p.adjust function}

\item{p.cut}{(numeric) the cutoff for multiple hypothesis corrected p.value}

\item{iter}{(numeric) Number of iterations to produce null distribution (note that null size will be iter*ncol(mat))}

\item{zerofreq}{(logical) whether to remove bins with observed frequency = 0; It is recommended to set to TRUE when the bins span genomic regions of low coverage}

\item{plot}{(logical) whether to generate a histogram comparing observed and null frequency distributions}

\item{verbose}{(logical) whether to return messages}
}
\value{
an instance of the class 'freq.cut'
}
\description{
Frequency hot spot detection
 
Obtains significance cutoff for the frequency of binary events encoded in a matrix such as that generated by shattered.regions and shattered.regions.cnv algorithms
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

## obtain a matrix of genomic bins vs samples indicating high density of breaks
shatt.regions <- shattered.regions.cnv(cnv)
mat <- shatt.regions@high.density.regions.hc

freq.p.test(mat)
}
\keyword{empirical}
\keyword{p.adjust}
\keyword{p.value,}
