% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.cnv.r
\name{gene.cnv}
\alias{gene.cnv}
\title{Gene-level CNV}
\usage{
gene.cnv(
  cnv,
  genome.v = "hg19",
  genesgr = NULL,
  chrlist = NULL,
  fill.gaps = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}

\item{genome.v}{(hg19 or hg38) reference genome version to fetch gene annotations when 'genesgr=NULL'}

\item{genesgr}{(S4) a GenomicRanges object containing genomic feature annotations (if not NULL overides genome.v). It must containg 'strand' and a metadata field 'gene_id' with unique values. Seqnames are expected in the format (chr1, chr2, ...)}

\item{chrlist}{(character) list of chromosomes to include chr1, chr2, etc...}

\item{fill.gaps}{(logical) whether to fill the gaps in the segmentation file using gap neighbour segmean average as log ratio}

\item{verbose}{(logical)}
}
\value{
an instance of the class 'genecnv' containing gene level copy number info
}
\description{
Obtains a gene-level copy number matrix from a segmentation profile.
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

gene.cnv(cnv)
}
\keyword{CNV,}
\keyword{genes}
\keyword{segmentation,}
