% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.cnv.r
\name{gene.cnv}
\alias{gene.cnv}
\title{Obtain a matrix with the weighted average CN per chromosome arm}
\usage{
gene.cnv(
  seg,
  genome.v = "hg19",
  chrlist = NULL,
  geneid = "Symbol",
  fill.gaps = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{genome.v}{(hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{chrlist}{(character) list of chromosomes to include chr1, chr2, etc...}

\item{geneid}{(character) only "Symbol" accepted; if NULL entrez ID will be used}
}
\description{
Obtain a matrix with the weighted average CN per chromosome arm
}
\examples{
gene.cnv()
}
\keyword{CNV,}
\keyword{genes}
\keyword{segmentation,}
