% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.track.view.r
\name{gene.track.view}
\alias{gene.track.view}
\title{Plot genome tracks in local genomic regions}
\usage{
gene.track.view(
  chr = NULL,
  start = NULL,
  stop = NULL,
  symbol = NULL,
  upstr = NULL,
  dnstr = NULL,
  genome.v = "hg19",
  cex.text = 0.6,
  addtext = TRUE,
  plot = TRUE,
  summary = TRUE,
  ...
)
}
\arguments{
\item{chr}{(character) chromosome (e.g chr9)}

\item{start}{(numeric) genomic coordinate from specified chromosome to start plotting}

\item{symbol}{(character) gene acceoted hgnc symbol to retrieve coordinates and area plotting ()}

\item{upstr}{(numeric) distance upstream specified gene to extend the area plotted}

\item{dnstr}{(numeric) distance downstream specified gene to extend the area plotted}

\item{genome.v}{(hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{addtext}{(logic)}

\item{plot}{(logic)}

\item{summary}{(logic) whether the function shoud return CNV segment 'segbrk' and SV 'svbrk' breakpoints tabular output}

\item{...}{additional plot parameters from graphics plot function}

\item{end}{(numeric) genomic coordinate from specified chromosome to stop plotting}

\item{addlegend}{(logic)}
}
\value{
a data.frame with gene isoform annotations and/or plot into open device
}
\description{
Plot genome tracks in local genomic regions
}
\examples{

gene.track.view(symbol="PTPRD", genome.v="hg19")
}
\keyword{CNV,}
\keyword{segmentation}
