% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.cnv.r
\docType{class}
\name{genecnv-class}
\alias{genecnv-class}
\alias{genecnv}
\title{Data class cnvmat}
\arguments{
\item{cnvmat}{(data.frame): matrix containing average CNV per gene (rows) for each sample (columns)}

\item{genesgr}{(S4): a GenomicRanges object with genomic feature annotations such as gene coordinates}

\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' validated by validate.cnv}

\item{param}{(list):}
}
\value{
an instance of the class 'genecnv' containing gene level copy number info
}
\description{
Class to store breakpoint annotations
}
