% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.genesgr.r
\name{get.genesgr}
\alias{get.genesgr}
\title{Genes GRanges}
\usage{
get.genesgr(genome.v = "hg19", chrlist = NULL)
}
\arguments{
\item{genome.v}{(hg19 or GRCh37 and hg38 or GRCh38) reference genome version to retrieve gene annotations}

\item{chrlist}{(character)}
}
\value{
a GRanges class object from the specified human genome version
}
\description{
Retrieves a GRanges object containinng gene annotations for an specified genome version
}
\examples{

get.genesgr(genome.v = "hg19",chrlist=NULL)

}
\keyword{CNV,}
\keyword{genes}
\keyword{segmentation,}
