% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot.spot.samples.R
\name{hot.spot.samples}
\alias{hot.spot.samples}
\title{Hot-spot sample retrieval}
\usage{
hot.spot.samples(chromo.regs.obj, freq.cut)
}
\arguments{
\item{chromo.regs.obj}{(chromo.regs) An object of class chromo.regs}

\item{freq.cut}{(numeric) the hot spot threshold above which peaks are defined for sample ID retrieval}
}
\value{
a list comprising two lists: peakRegions, peakRegionsSamples
}
\description{
Collects sample ids with shattered regions detected at hot-spots based on certain p-value cutoff
}
\examples{
# validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)
svc <- validate.svc(svdat_lung_ccle)

chromo.regs.obj <- shattered.regions(cnv,svc)
mat<-hbd.mat(chromo.regs.obj)

pcut.obj <- freq.p.test(mat,plot=FALSE)
pcut <- freq.threshold(pcut.obj)

res <- hot.spot.samples(chromo.regs.obj,pcut)

}
