% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot.spot.samples.R
\name{hot.spot.samples}
\alias{hot.spot.samples}
\title{Define peak regions and samples in shattered regions hot spots}
\usage{
hot.spot.samples(chromo.regs, freq.cut)
}
\arguments{
\item{chromo.regs}{(list) an object returned by 'shattered.regions' and 'shattered.regions.cnv'}

\item{freq.cut}{(numeric) the hot spot threshold above which peaks are defined and sample ids collected}
}
\description{
Define peak regions and samples in shattered regions hot spots
}
