% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{IQSD}
\alias{IQSD}
\title{Inter-quantile standard deviation}
\usage{
IQSD(x, lowQ = 0.1, upQ = 0.9)
}
\arguments{
\item{x}{numeric vector to compute interquantile standard deviation}

\item{lowQ}{lower quantile}

\item{upQ}{upper quantile}
}
\value{
(numeric) the IQSD value
}
\description{
Obtains inter quantile standard deviation for a defined 'x' vector and both lower and upper quantiles
}
\examples{

x <- rnorm(100)
IQSD(x)
}
\keyword{interquartile}
\keyword{statistics,}
