% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{map2color}
\alias{map2color}
\title{Color map from numeric vector}
\usage{
map2color(x, pal = NULL, limits = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{pal}{color palette}

\item{limits}{numeric limit fr color mapping}
}
\value{
a color vector graded according to x
}
\description{
Produces a vector of colors based on a given palette. The colors are defined by the inpuit vector
}
\examples{

x <- rnorm(100)
x_color <- map2color(x)
head(x_color)
}
\keyword{color,}
\keyword{number}
