% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.density.r
\name{match.breaks}
\alias{match.breaks}
\title{Breakpoint matching}
\usage{
match.breaks(brk1, brk2, maxgap = 1e+05, verbose = FALSE, plot = TRUE)
}
\arguments{
\item{brk1}{(S4) an object of class breaks as returned by `svc.breaks` and `cnv.breaks`}

\item{brk2}{(S4) an object of class breaks as returned by `svc.breaks` and `cnv.breaks` to compare against brk1}

\item{maxgap}{(numeric) distance (base pairs) limit for nreakpoints to be consider colocalized}

\item{verbose}{(logical) whether to return internal messages}

\item{plot}{(logical) whether to plot into open device}
}
\value{
an object containing co-localizing breakpoints from two input 'breaks'
}
\description{
Match common breakpoints from two different datasets or data types based on their co-localization in the genome.
}
\examples{

# initialize CNV and SVC data
cnv <- validate.cnv(segdat_lung_ccle)
svc <- validate.svc(svdat_lung_ccle)

## Obtain breakpoints from CNV and SVC
brk1 <- cnv.breaks(cnv)
brk2 <- svc.breaks(svc)

common.brk <- match.breaks(brk1, brk2)

}
\keyword{CNV,}
\keyword{SV,}
\keyword{breakpoints}
\keyword{genomic}
