% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.breaks.r
\name{match.breaks}
\alias{match.breaks}
\title{Identify CNV breakpoints provided a segmentation file}
\usage{
match.breaks(brk1, brk2, maxgap = 50000, verbose = FALSE)
}
\arguments{
\item{brk1}{(data.frame) brakpoint coordinates 1 (sample, chrom, pos)}

\item{brk2}{(data.frame) brakpoint coordinates 1 (sample, chrom, pos)}

\item{verbose}{(logical)}
}
\description{
Identify CNV breakpoints provided a segmentation file
}
\examples{
match.breaks()
}
\keyword{CNV,}
\keyword{SV,}
\keyword{breakpoints}
\keyword{genomic}
