% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.density.r
\name{match.breaks}
\alias{match.breaks}
\title{Compare two different sets of breakpoints}
\usage{
match.breaks(brk1, brk2, maxgap = 1e+05, verbose = FALSE, plot = TRUE)
}
\arguments{
\item{brk1}{(S4) an object of class breaks as returned by `svc.breaks` or `cnv.breaks`}

\item{brk2}{(S4) an object of class breaks as returned by `svc.breaks` or `cnv.breaks` to compare against brk1}

\item{maxgap}{(numeric) distance (base pairs) limit for breakpoints to be considered colocalized}

\item{verbose}{(logical)}
}
\value{
an object containing co-localizing breakpoints from two input 'breaks'
}
\description{
Compare two different sets of breakpoints
}
\examples{

## Obtain breakpoints from segmentation data
cnv <- validate.cnv(segdat_lung_ccle)
brk1 <- cnv.breaks(cnv)

## Obtain breakpoints from SV calls data
sv <- validate.svc(svdat_lung_ccle)
brk2 <- svc.breaks(svc)

common.brk <- match.breaks(brk1, brk2)

## average percentage of colocalizing breaks
restab <- data.frame(common.breaks$restab)[order(common.breaks$restab$total.brk2),]
m2 <- sprintf("\%.1f",100*mean(restab$matched.brk2/restab$total.brk2)) 

## Plot the proportion of SV breakpoints that have colocalizing CNV breakpoints
barplot(rbind(restab$matched.brk2, restab$total.brk2 - restab$matched.brk2),
        border=NA,las=2,xlab="",horiz=FALSE,cex.main=.7,cex.names=.4, names=rownames(restab))
legend("top",paste("SV breaks matched by CNV breaks\n","Average = ",m2,"\%",sep=""),bty='n')
grid(ny=NULL,nx=NA)
}
\keyword{CNV,}
\keyword{SV,}
\keyword{breakpoints}
\keyword{genomic}
