% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.breaks.r
\name{match.breaks}
\alias{match.breaks}
\title{Identify CNV breakpoints provided a segmentation file}
\usage{
match.breaks(brk1, brk2, maxgap = 50000, verbose = FALSE)
}
\arguments{
\item{brk1}{(data.frame) brakpoint coordinates 1 (sample, chrom, pos), as returned by `sv.breaks` and `seg.breaks`}

\item{brk2}{(data.frame) brakpoint coordinates 1 (sample, chrom, pos), as returned by `sv.breaks` and `seg.breaks`}

\item{maxgap}{(numeric) distance (base pairs) limit for nreakpoints to be consider colocalized}

\item{verbose}{(logical)}
}
\description{
Identify CNV breakpoints provided a segmentation file
}
\examples{

## Obtain breakpoints from segmentation data
seg <- validate.seg(segdat_lung_ccle)
brk1 <- seg.breaks(seg)

## Obtain breakpoints from SV calls data
sv <- validate.sv(svdat_lung_ccle)
brk2 <- sv.breaks(sv)

common.brk <- match.breaks(brk1, brk2)

## average percentage of colocalizing breaks
restab <- data.frame(common.breaks$restab)[order(common.breaks$restab$total.brk2),]
m2 <- sprintf("\%.1f",100*mean(restab$matched.brk2/restab$total.brk2)) 

## Plot the proportion of SV breakpoints that have colocalizing CNV breakpoints
barplot(rbind(restab$matched.brk2, restab$total.brk2 - restab$matched.brk2),
        border=NA,las=2,xlab="",horiz=FALSE,cex.main=.7,cex.names=.4, names=rownames(restab))
legend("top",paste("SV breaks matched by CNV breaks\n","Average = ",m2,"\%",sep=""),bty='n')
grid(ny=NULL,nx=NA)
}
\keyword{CNV,}
\keyword{SV,}
\keyword{breakpoints}
\keyword{genomic}
