% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.means.r
\name{med.segmean}
\alias{med.segmean}
\title{Median sample CNV}
\usage{
med.segmean(cnv)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}
}
\value{
(numeric) a vector containing the median logR value of a segmented data.frame
}
\description{
Obtain the median weighted segment mean from a segmentaton file; The weighted median refers to the logR that occupies a center of all segments ordered by their log ratio
}
\examples{

## validate input CNV data.frames
cnv <- validate.cnv(segdat_lung_ccle)

med_seg_mean <- med.segmean(cnv)
head(med_seg_mean)
}
\keyword{CNV,}
\keyword{segmentation}
