% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.segmean.r
\name{med.segmean}
\alias{med.segmean}
\title{Obtain the median segment mean from a segmentaton file}
\usage{
med.segmean(seg)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\description{
Obtain the median segment mean from a segmentaton file
}
\examples{
med.segmean()
}
\keyword{CNV,}
\keyword{segmentation}
