% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.means.r
\name{med.segmean}
\alias{med.segmean}
\title{Obtain the median segment mean from a segmentaton file}
\usage{
med.segmean(cnv)
}
\arguments{
\item{cnv}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\description{
Obtain the median segment mean from a segmentaton file
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

med.segmean(cnv)
}
\keyword{CNV,}
\keyword{segmentation}
