% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{merge2lists}
\alias{merge2lists}
\title{Merge two lists}
\usage{
merge2lists(x, y, fun = "unique")
}
\arguments{
\item{x}{(list): input list 1}

\item{y}{(list): input list 2}

\item{fun}{(character): Either 'unique' or 'intersect' are accepted}
}
\value{
(list) merged list from x and y
}
\description{
Merge of 2 lists into one that contains unique or intersect vectors for each list entry with shared names
}
\examples{

x <- sapply(letters[1:10], function(i) sample(1:10)[1:sample(2:10)[1]], simplify=FALSE )
y <- sapply(letters[5:15], function(i) sample(1:10)[1:sample(2:10)[1]], simplify=FALSE )
merge2lists(x,y)
}
\keyword{lists}
\keyword{merge}
