% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge2lists.r
\name{merge2lists}
\alias{merge2lists}
\title{Given two lists with (or without common names) returns a combined list; if common names their values are returned unique}
\usage{
merge2lists(x, y)
}
\arguments{
\item{x}{list 1}

\item{y}{list 2}
}
\value{
a combined list
}
\description{
Given two lists with (or without common names) returns a combined list; if common names their values are returned unique
}
\examples{
merge2lists(x,y)
}
\keyword{list}
