% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.r
\name{merge2lists}
\alias{merge2lists}
\title{Given two lists with (or without common names) returns a combined list; if common names their values merged and returned as unique}
\usage{
merge2lists(x, y, fun = "unique")
}
\arguments{
\item{x}{(list) input list 1}

\item{y}{(list) input list 2}

\item{func}{(character) Either 'unique' or 'intersect' are accepted}
}
\value{
a combined list
}
\description{
Given two lists with (or without common names) returns a combined list; if common names their values merged and returned as unique
}
\examples{

x <- sapply(letters[1:10], function(i) sample(1:10)[1:sample(2:10)[1]], simplify=FALSE )
y <- sapply(letters[5:15], function(i) sample(1:10)[1:sample(2:10)[1]], simplify=FALSE )
merge2lists(x,y)
}
\keyword{list}
