% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.genome.changed.r
\name{pct.genome.changed}
\alias{pct.genome.changed}
\title{Calculate the percentage of genome changed by CNV}
\usage{
pct.genome.changed(seg, ch.pct = 0.2, discard.sex = TRUE)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{ch.pct}{(numeric) percentage CNV gain/loss for a segment to be considered changed (i.e. 0.2 = 20 percent change 0.8 < segmean && segmean > 1.2)}

\item{discard.sex}{(logical) whether sex chromosomes should be included or not}
}
\description{
Calculate the percentage of genome changed by CNV
}
\examples{
break.density()
}
\keyword{CNV,}
\keyword{segmentation}
