% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{data}
\name{refseq_hg19}
\alias{refseq_hg19}
\title{Reference transcript and exon annotations for hg19}
\format{
An object of class \code{refSeqDat} of length 1.
}
\usage{
refseq_hg19
}
\description{
refSeq annotations for hg19 version from UCSC (http://genome.ucsc.edu/cgi-bin/hgTables)
}
\keyword{exons}
\keyword{genes,}
\keyword{transcripts,}
