% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpluscnv.data.r
\docType{data}
\name{segdat_lung_ccle}
\alias{segdat_lung_ccle}
\title{Lung CCLE CNV data}
\format{An object of class \code{data.frame} with 134358 rows and 6 columns.}
\usage{
segdat_lung_ccle
}
\description{
CCLE CNV segmentation data from LUNG tissue cell lines (DepMap): https://depmap.org/portal/download/
}
\keyword{CNV}
\keyword{segmentation}
