% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcnvplus.data.r
\docType{data}
\name{segdat_lung_ccle}
\alias{segdat_lung_ccle}
\title{CCLE degmentation data from breast tissue cell lines}
\format{An object of class \code{data.frame} with 162800 rows and 6 columns.}
\usage{
data()
}
\description{
Affy SNP6.0 array based segmentation data for lung cancer (DepMap): https://depmap.org/portal/download/
TARGET CGI structural variants and CNV segmentation: https://target-data.nci.nih.gov/
}
\examples{
data(segdat_lung_ccle)
data(svdat_lung_ccle)
data(cnv_blacklist_regions)
data(nbl_segdat)
data(nbl_svdat)
}
\keyword{CNV}
\keyword{SVs}
\keyword{segmentation,}
