% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.cnv.artifact.r
\name{segment.gap}
\alias{segment.gap}
\title{CNV segmentation gap filling}
\usage{
segment.gap(cnv, minsize = 5000, chrlist = NULL, verbose = FALSE)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}

\item{minsize}{(numeric) the minimum gap size required to fill the gap}

\item{chrlist}{(character) list of chromosomes to include chr1, chr2, etc...}

\item{verbose}{(logical) whether to return internal messages}
}
\value{
a data.frame containing CNV data
}
\description{
Fills the gaps in a segmentation data.frame. Chromosome limits are defined for the complete segmentation dataset then segments fill the missing terminal regions. 
The CN log-ratio of the added segments is set to the average of the closest neighbours in each sample.
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)

cnv2 <- segment.gap(cnv)
cnv2
}
\keyword{CNV,}
\keyword{segmentation}
