% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.gap.r
\name{segment.gap}
\alias{segment.gap}
\title{Filling gaps in a segmentation dat abased on chromosome limits defined withing the complete segmentation dataset}
\usage{
segment.gap(seg, chrlist = NULL, verbose = TRUE)
}
\arguments{
\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{chrlist}{(character) list of chromosomes to include chr1, chr2, etc...}
}
\description{
Filling gaps in a segmentation dat abased on chromosome limits defined withing the complete segmentation dataset
}
\examples{
segment.gap()
}
\keyword{CNV,}
\keyword{segmentation}
