% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.r
\name{shattered.eval}
\alias{shattered.eval}
\title{Evaluate true catastrophic events
Evaluate shattered regions based on interleaved breaks and breakpoint dispersion parameters in order to identify true catastrophic chromosomal alterations}
\usage{
shattered.eval(
  chromo.regs.obj,
  interleaved.cut = 0.5,
  dist.iqm.cut = 1e+05,
  verbose = TRUE
)
}
\arguments{
\item{chromo.regs.obj}{(chromo.regs) An object of class chromo.regs}

\item{interleaved.cut}{(numeric) the percentage of non interleaved structural variant calls}

\item{dist.iqm.cut}{(numeric) interquantile average of the distance between breakpoints within a shattered region}

\item{verbose}{(logical)}
}
\value{
an instance of the class 'chromo.regs' containing breakpoint mapping onto genes
}
\description{
Evaluate true catastrophic events
Evaluate shattered regions based on interleaved breaks and breakpoint dispersion parameters in order to identify true catastrophic chromosomal alterations
}
