% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.eval.r
\name{shattered.eval}
\alias{shattered.eval}
\title{Evaluate shattered regions}
\usage{
shattered.eval(
  chromo.regs,
  interleaved.cut = 0.5,
  dist.iqm.cut = 1e+05,
  verbose = TRUE
)
}
\arguments{
\item{chromo.regs}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{interleaved.cut}{(numeric) the percentage of non interleaved structural variant calls}

\item{dispersion.cut}{(numeric) a measure of how disperse are SV and CNV breakpoints in a shattered region;}
}
\description{
Evaluate shattered regions
}
\examples{
shattered.eval()
}
\keyword{chromothripsis,}
\keyword{segmentation,}
\keyword{structural}
\keyword{variants}
