% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.eval.r
\name{shattered.eval}
\alias{shattered.eval}
\title{Evaluate shattered regions based on interleaved breaks and breakpoint dispersion parameters}
\usage{
shattered.eval(
  chromo.regs,
  interleaved.cut = 0.5,
  dist.iqm.cut = 1e+05,
  verbose = TRUE
)
}
\arguments{
\item{chromo.regs}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{interleaved.cut}{(numeric) the percentage of non interleaved structural variant calls}

\item{dist.iqm.cut}{(numeric) interquantile average of the distance between breakpoints within a shattered region}

\item{verbose}{(logical)}
}
\description{
Evaluate shattered regions based on interleaved breaks and breakpoint dispersion parameters
}
