% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.map.plot.r
\name{shattered.map.plot}
\alias{shattered.map.plot}
\title{Shattered regions genomic map}
\usage{
shattered.map.plot(
  chromo.regs.obj,
  conf = "hc",
  genome.v = "hg19",
  chrlist = NULL,
  freq.cut = NULL,
  add.legend = "top"
)
}
\arguments{
\item{chromo.regs.obj}{(chromo.regs) An object of class chromo.regs}

\item{conf}{(character) either 'hc' for high confidence objects or else all included}

\item{genome.v}{(character)  reference genome version to draw chromosome limits and centromeres either hg19 or hg38 accepted}

\item{chrlist}{(character) vector containing chromosomes to include in the analysis; if NULL all chromosomes available in the input will be included}

\item{freq.cut}{the value to draw an horizontal line; use 'freq.p.test' to obtain a threshold for statisticaly significant hot spots}

\item{add.legend}{the position of the legend in the plot; if null, no legend will be draw}
}
\value{
a plot into open device
}
\description{
Plots a genome wide map of shattered region frequencies
}
\examples{


## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)
svc <- validate.svc(svdat_lung_ccle)

## obtain shattered regions
chromo.regs.obj <- shattered.regions(cnv,svc)

shattered.map.plot(chromo.regs.obj)
}
\keyword{chromosome}
\keyword{genome}
\keyword{map}
\keyword{shattering,}
