% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.map.plot.r
\name{shattered.map.plot}
\alias{shattered.map.plot}
\title{Plot CNV frequency across the human genome from a seg file sontaining multiple samples}
\usage{
shattered.map.plot(
  shatt_regs,
  conf = "hc",
  hg = "hg19",
  fdr_cut = NULL,
  add.legend = "top"
)
}
\arguments{
\item{shatt_regs}{object returned by 'shattered.regions' or 'shattered.regions.cnv'}

\item{conf}{(character) either 'hc' for high confidence objects or else all included}

\item{hg}{(hg19 or h38) reference genome version to draw chromosome limits and centromeres}

\item{fdr_cut}{the value to draw an horizontal line based}

\item{add.legend}{the position of the legend in the plot; if null, no legend will be draw}
}
\description{
Plot CNV frequency across the human genome from a seg file sontaining multiple samples
}
\examples{
shattered.map.plot()
}
\keyword{chromosome}
\keyword{genome}
\keyword{map}
\keyword{shattering,}
