% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.map.plot.r
\name{shattered.map.plot}
\alias{shattered.map.plot}
\title{Plot CNV frequency across the human genome from a seg file sontaining multiple samples}
\usage{
shattered.map.plot(
  shatt.regions,
  conf = "hc",
  genome.v = "hg19",
  freq.cut = NULL,
  add.legend = "top"
)
}
\arguments{
\item{shatt.regions}{object returned by 'shattered.regions' or 'shattered.regions.cnv'}

\item{conf}{(character) either 'hc' for high confidence objects or else all included}

\item{genome.v}{(hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{freq.cut}{the value to draw an horizontal line; use 'freq.p.test' to obtain a threshold for statisticaly significant hot spots}

\item{add.legend}{the position of the legend in the plot; if null, no legend will be draw}
}
\description{
Plot CNV frequency across the human genome from a seg file sontaining multiple samples
}
\examples{


require(circlize)

## validate input data.frames
seg <- validate.seg(segdat_lung_ccle)
sv <- validate.sv(svdat_lung_ccle)

## obtain shattered regions
shatt.regions <- shattered.regions(seg,sv)

shattered.map.plot(shatt.regions)
}
\keyword{chromosome}
\keyword{genome}
\keyword{map}
\keyword{shattering,}
