% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shattered.regions.r
\name{shattered.regions}
\alias{shattered.regions}
\title{Shattered region detection}
\usage{
shattered.regions(
  cnv,
  svc,
  fc.pct = 0.2,
  min.cnv.size = 0,
  min.num.probes = 0,
  low.cov = NULL,
  clean.brk = NULL,
  window.size = 10,
  slide.size = 2,
  num.cnv.breaks = 6,
  num.cnv.sd = 5,
  num.svc.breaks = 6,
  num.svc.sd = 5,
  num.common.breaks = 3,
  num.common.sd = 3,
  maxgap = 10000,
  chrlist = NULL,
  interleaved.cut = 0.5,
  dist.iqm.cut = 1e+05,
  verbose = TRUE
)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}

\item{svc}{(S4) an object of class svcnvio containing data type 'svc' initialized by validate.svc}

\item{fc.pct}{(numeric) inherited from cnv.breaks(); copy number change between 2 consecutive segments: i.e (default) cutoff = 0.2 represents a fold change of 0.8 or 1.2}

\item{min.cnv.size}{(numeric) inherited from cnv.breaks(); The minimun segment size (in base pairs) to include in the analysis}

\item{min.num.probes}{(numeric) inherited from cnv.breaks(); The minimun number of probes per segment to include in the analysis}

\item{low.cov}{(data.frame) inherited from cnv.breaks(), svc.breaks() and match.breaks; a data.frame (chr, start, end) indicating low coverage regions to exclude from the analysis}

\item{clean.brk}{(numeric) inherited from cnv.breaks(); n cutoff for redundant breakpoints to filter out; if NULL, no filter will be applied}

\item{window.size}{(numeric) size in megabases of the genmome bin to compute break density}

\item{slide.size}{(numeric) size in megabases of the sliding genmome window}

\item{num.cnv.breaks}{(numeric) number of segmentation breakpoints per segments to be considered high-density break}

\item{num.cnv.sd}{(numeric) number of standard deviations above the sample average for num.cnv.breaks}

\item{num.svc.breaks}{(numeric) number of svc breakpoints per segments to be considered high-density break}

\item{num.svc.sd}{(numeric) number of standard deviations above the sample average for num.svc.breaks}

\item{num.common.breaks}{(numeric) number of common SV and segmentation breakpoints per segments to be considered high-density break}

\item{num.common.sd}{(numeric) number of standard deviations above the sample average for num.common.breaks}

\item{maxgap}{(numeric) inherited from match.breaks(); sets the maximum gap between co-localizing orthogonal breakpoints}

\item{chrlist}{(character) vector containing chromosomes to include in the analysis; if NULL all chromosomes available in the input will be included}

\item{interleaved.cut}{(numeric) 0-1 value indicating percentage of interleaved (non-contiguous) SV breakpoint pairs}

\item{dist.iqm.cut}{(numeric) interquantile average of the distance between breakpoints within a shattered region}

\item{verbose}{(logical)}
}
\value{
an instance of the class 'chromo.regs' containing breakpoint mapping onto genes
}
\description{
Caller for the identification of shattered genomic regions based on CNV and SVC data
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)
svc <- validate.svc(svdat_lung_ccle)

shattered.regions(cnv,svc)
}
\keyword{chromoplexy,}
\keyword{chromosome}
\keyword{chromothripsis,}
\keyword{shattering}
