% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.break.annot.r
\name{sv.break.annot}
\alias{sv.break.annot}
\title{Identify recurrently altered genes}
\usage{
sv.break.annot(
  svdf,
  genome.v = "hg19",
  maxgap = 0,
  upstr = 50000,
  sv.seg.size = 2e+05,
  verbose = TRUE
)
}
\arguments{
\item{svdf}{(data.frame) structural variant table including  8 columns: sample, chrom1, pos1, strand1, chrom2, pos2, strand2, svclass}

\item{genome.v}{(hg19 or hg38) reference genome version to draw chromosome limits and centromeres}

\item{maxgap}{(numeric) offset region size (base pairs) beyond transcription start and end sites}

\item{upstr}{(numeric) upstream region size (base pairs) for searching breakpoints in gene regulatory regions}

\item{sv.seg.size}{(numeric) segmental variants (DEL, DUP, INV or INS) size limit; bigger sigments are considered translocations}
}
\description{
Identify recurrently altered genes
}
\examples{
sv.break.annot()
}
\keyword{Structural}
\keyword{annotation}
\keyword{variants}
