% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.model.view.r
\name{sv.model.view}
\alias{sv.model.view}
\title{SV integrated visualization}
\usage{
sv.model.view(
  cnv,
  svc,
  chrom,
  start,
  stop,
  sampleids = NULL,
  cnvlim = c(-2, 2),
  addlegend = "both",
  cex.legend = 1,
  interval = NULL,
  addtext = NULL,
  cex.text = 0.8,
  plot = TRUE,
  summary = TRUE,
  ...
)
}
\arguments{
\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' initialized by validate.cnv}

\item{svc}{(S4) an object of class svcnvio containing data type 'svc' initialized by validate.svc}

\item{chrom}{(character) chromosome (e.g chr9)}

\item{start}{(numeric) genomic coordinate from specified chromosome to start plotting}

\item{stop}{(numeric) genomic coordinate from specified chromosome to stop plotting}

\item{sampleids}{(character) a vector containing a list of sample ids represented in svc and/or cnv objects to be plotted}

\item{cnvlim}{(numeric) limits for color coding of background CNV log-ratios. Use to modify the CNV color contrast at different levels.}

\item{addlegend}{(character) One of 'sv' (show SV type legend), 'cnv' (show CNV background color legend) or 'both'.}

\item{cex.legend}{(numeric) The cex values for each legend}

\item{interval}{(numeric) The axis interval in base pairs}

\item{addtext}{(character) a vector indicating what SV types should include text labels indicating brakpoint partners genomic locations. The added labels are point breakpoint locations outside the plot area. (e.g. c("TRA","INV") )}

\item{cex.text}{(numeric) The magnification to be used for SV text info added}

\item{plot}{(logic) whether to produce a graphical output}

\item{summary}{(logic) whether the function shoud return CNV segment 'segbrk' and SV 'svbrk' breakpoints tabular output}

\item{...}{additional plot parameters from graphics plot function}
}
\value{
a data.frame with CNV and SVN breakpoint annotations and/or plot into open device
}
\description{
Integrated visualization of SVC and CNV data for defined genomic locations. CNV and SVC data is overlayed into a sample-based track visualization map.
}
\examples{

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)
svc <- validate.svc(svdat_lung_ccle)

# obtain the coordinates of a desired genomic regionbased on a known gene locus 
refSeqGene <- gene.symbol.info(refseq_hg19,"PTPRD")
start <- refSeqGene$start - 150000;
stop <- refSeqGene$stop+ 50000;
chrom <- refSeqGene$chrom

sv.model.view(cnv, svc, chrom, start, stop)

}
\keyword{CNV,}
\keyword{segmentation}
\keyword{structural}
\keyword{variant,}
