% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.model.view.r
\name{sv.model.view}
\alias{sv.model.view}
\title{Integrated visualization of SVs and CNVs in local genomic regions}
\usage{
sv.model.view(
  svc,
  cnv,
  chr,
  start,
  stop,
  sampleids = NULL,
  cnvlim = c(-2, 2),
  addlegend = "both",
  cex.legend = 1,
  interval = NULL,
  addtext = NULL,
  cex.text = 0.8,
  plot = TRUE,
  summary = TRUE,
  ...
)
}
\arguments{
\item{svc}{(S4) an object of class svcnvio containing data type 'svc' validated by validate.svc}

\item{cnv}{(S4) an object of class svcnvio containing data type 'cnv' validated by validate.cnv}

\item{chr}{(character) chromosome (e.g. chr9)}

\item{start}{(numeric) genomic coordinate from specified chromosome to start plotting}

\item{stop}{(numeric) genomic coordinate from specified chromosome to stop plotting}

\item{sampleids}{(character) vector containing list of samples to include in plot. if set to NULL, all samples in the input will be used}

\item{cnvlim}{(numeric) limits for color coding of background CNV log-ratios. Use to modify the CNV color contrast at different levels.}

\item{addlegend}{(character) one of 'sv' (show SV type legend), 'cnv' (show CNV background color legend) or 'both'.}

\item{cex.legend}{(numeric) the cex values for each legend}

\item{addtext}{(character) a vector indicating what SV types should include text labels indicating breakpoint partners' genomic locations. 
The added labels are breakpoint locations outside the plot area. (e.g. c("TRA","INV") )}

\item{cex.text}{(numeric) the cex value for text}

\item{plot}{(logic) whether to produce a graphical output}

\item{summary}{(logic) whether the function shoud return CNV segment 'segbrk' and SV 'svbrk' breakpoints tabular output}

\item{...}{additional plot parameters from graphics plot function}
}
\value{
a data.frame with CNV and SVN breakpoint annotations and/or plot into open device
}
\description{
Integrated visualization of SVs and CNVs in local genomic regions
}
\examples{
plot()

## validate input data.frames
cnv <- validate.cnv(segdat_lung_ccle)
svc <- validate.svc(svdat_lung_ccle)

# obtain the coordinates of a desired genomic regionbased on a known gene locus 
df <- gene.track.view(symbol = "PTPRD", plot=FALSE)$df
start <- min(df$txStart) - 150000;
stop <- max(df$txEnd) + 50000;
chr <- df$chrom[1]

sv.model.view(svc, cnv, chr, start, stop,addlegend = TRUE, addtext=c("TRA"))

}
\keyword{CNV,}
\keyword{segmentation}
