% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv.model.view.r
\name{sv.model.view}
\alias{sv.model.view}
\title{Integrated visualization of SVs and SNV in local genomic regions}
\usage{
sv.model.view(
  sv,
  seg,
  chr,
  start,
  stop,
  sampleids = NULL,
  cnvlim = c(-2, 2),
  addlegend = "both",
  cex.legend = 1,
  interval = NULL,
  addtext = NULL,
  cex.text = 0.8,
  plot = TRUE,
  summary = TRUE,
  ...
)
}
\arguments{
\item{sv}{(data.frame) structural variant table including  8 columns: sample, chrom1, pos1, strand1, chrom2, pos2, strand2, svclass}

\item{seg}{(data.frame) segmentation data with 6 columns: sample, chromosome, start, end, probes, segment_mean}

\item{chr}{(character) chromosome (e.g chr9)}

\item{start}{(numeric) genomic coordinate from specified chromosome to start plotting}

\item{sampleids}{(character)}

\item{cnvlim}{(numeric) limits for color coding of background CNV log-ratios. Use to modify the CNV color contrast at different levels.}

\item{addlegend}{(character) one of 'sv' (show SV type legend), 'cnv' (show CNV background color legend) or 'both'.}

\item{cex.legend}{(numeric) the cex values for each legend}

\item{addtext}{(character) a vector indicating what SV types should include text labels indicating brakpoint partners genomic locations. 
The labeels are only added to point tbreakpoint locations outside the plot area. (e.g. c("TRA","INV") )}

\item{plot}{(logic) whether to produce a graphical output}

\item{summary}{(logic) whether the function shoud return CNV segment 'segbrk' and SV 'svbrk' breakpoints tabular output}

\item{...}{additional plot parameters from graphics plot function}

\item{end}{(numeric) genomic coordinate from specified chromosome to stop plotting}
}
\description{
Integrated visualization of SVs and SNV in local genomic regions
}
\examples{
plot()

## validate input data.frames
seg <- validate.seg(segdat_lung_ccle)
sv <- validate.sv(svdat_lung_ccle)

# obtain the coordinates of a desired genomic regionbased on a known gene locus 
df <- gene.track.view(symbol = "PTPRD", plot=FALSE)$df
start <- min(df$txStart) - 150000;
stop <- max(df$txEnd) + 50000;
chr <- df$chrom[1]

sv.model.view(sv, seg, chr, start, stop,addlegend = TRUE, addtext=c("TRA"))

}
\keyword{CNV,}
\keyword{segmentation}
