% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoint.density.r
\name{svc.breaks}
\alias{svc.breaks}
\title{Identify SVC breakpoints}
\usage{
svc.breaks(svc, chrlist = NULL, low.cov = NULL)
}
\arguments{
\item{svc}{(S4) an object of class svcnvio containing data type 'svc' initialized by validate.svc}

\item{chrlist}{(character) list of chromosomes to include chr1, chr2, etc...}

\item{low.cov}{(data.table) a data.table (chrom, start, end) indicating low coverage regions to exclude from the analysis}
}
\value{
an instance of the class 'breaks' containing breakpoint and breakpoint burden information
}
\description{
Transform structural varian (SVC) data.frame into a 'breaks' object
}
\examples{

## Obtain breakpoints from SV calls data
svc <- validate.svc(svdat_lung_ccle)

svc.breaks(svc)
}
\keyword{Structural}
\keyword{variants}
