% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.data.r
\docType{class}
\name{svcnvio-class}
\alias{svcnvio-class}
\alias{svcnvio}
\title{Data class svcnvio}
\arguments{
\item{data}{(data.table): cnv or svc data.table to be validated by 'validate.cnv' or 'validate.svc' respectivelly}

\item{type}{(character): the data type  "cnv" or "svc" defined by "validate.cnv" or "validate.svc" respectivelly}
}
\value{
an instance of the class 'svcnvio' containing SV data derived from CNV or SVC data types;  A unique id (uid) column is also added
}
\description{
Class to store CNV segmentation data
}
\seealso{
Additional data format information in the man pages of validate.cnv and validate.svc
}
