% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.data.r
\name{validate.cnv}
\alias{validate.cnv}
\title{Initialization of CNV data}
\usage{
validate.cnv(cnv.df)
}
\arguments{
\item{cnv.df}{(data.frame) segmentation data with at least 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\value{
an instance of the class 'svcnvio' containing segmentation data derived from CNV data type;  A unique id (uid) column is also added
}
\description{
This function validates and reformats the CNV segmentation data type containing copy number log-ratios. It is used internaly by 'svpluscnv' functions that require this type of data.
}
\examples{

validate.cnv(segdat_lung_ccle)
}
\keyword{CNV,}
\keyword{segmentation}
