% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.data.r
\name{validate.cnv}
\alias{validate.cnv}
\title{Validate and reformat CNV segmentation input}
\usage{
validate.cnv(cnv.df)
}
\arguments{
\item{cnv.df}{(data.frame) segmentation data with at least 6 columns: sample, chromosome, start, end, probes, segment_mean}
}
\value{
(data.frame) CNV segmentation table with validated format for other svpluscnv operations;  A unique id (uid) column is also added
}
\description{
Validate and reformat CNV segmentation input
}
\examples{
validate.cnv()
}
\keyword{CNV,}
\keyword{segmentation}
