% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.sv.r
\name{validate.sv}
\alias{validate.sv}
\title{This function calculates the percent genome changed by CNV
sv classes accepted: DEL(deletion), DUP(duplication), INS(insertion), TRA(translocation), INV(inversion)}
\usage{
validate.sv(sv)
}
\arguments{
\item{sv}{(data.frame) structural variant table including  8 columns: sample, chrom1, pos1, strand1, chrom2, pos2, strand2, svclass}
}
\description{
This function calculates the percent genome changed by CNV
sv classes accepted: DEL(deletion), DUP(duplication), INS(insertion), TRA(translocation), INV(inversion)
}
\examples{
validate.sv()
}
\keyword{SV,}
\keyword{structural}
\keyword{variants}
